Item and Test Information Function

This function computes item and test information functions (Hambleton et al., 1991) for a given set of theta values.

Usage

info(x, ...)

# Default S3 method
info(x, theta, D = 1, tif = TRUE, ...)

# S3 method for class 'est_item'
info(x, theta, tif = TRUE, ...)

# S3 method for class 'est_irt'
info(x, theta, tif = TRUE, ...)

Arguments

x

A data frame containing item metadata (e.g., item parameters, number of categories, IRT model types, etc.); or an object of class est_irt obtained from est_irt(), or est_item from est_item().

See est_irt() or simdat() for more details about the item metadata. This data frame can be easily created using the shape_df() function.

...

Further arguments passed to or from other methods.

theta

A numeric vector of theta values at which item and test information are computed.

D

A scaling constant used in IRT models to make the logistic function closely approximate the normal ogive function. A value of 1.7 is commonly used for this purpose. Default is 1.

tif

Logical. If TRUE, the test information function is computed. Default is TRUE.

Value

This function returns an object of class info, which is a list containing the following components:

iif

A matrix of item information values. Each row corresponds to an item, and each column represents the information value computed at a given theta point. The row names are the item IDs, and the column names indicate the theta points (e.g., "theta.1", "theta.2", ...).

tif

A numeric vector containing the test information values at each theta value, computed as the sum of item information values across all items. This component is included only when tif = TRUE.

theta

A numeric vector of theta values at which the item and test information functions are evaluated. This matches the user-supplied theta argument.

The returned object is of class info and can be visualized using the function plot.info(). This output structure is consistent across input types (data.frame, est_item, est_irt), and facilitates downstream plotting, comparison, or export of information function values.

Details

This function calculates the amount of statistical information provided by each item (item information function, IIF) and the total test (test information function, TIF) across a range of ability (theta) values. Higher information values at a particular theta level indicate greater measurement precision at that ability level.

The input x must follow a specific data frame format if not already an est_irt or est_item object. The structure of this data frame is explained in the documentation of est_irt() and simdat(). Items of different models (e.g., 3PLM, GPCM) can be combined in a single test.

The information is computed for each item appropriately and aggregated for the TIF if tif = TRUE. The TIF is often used to assess where the test provides the most precision, and is critical when designing adaptive tests or evaluating test coverage across the ability continuum.

The returned object is a list of class "info", which contains the item information matrix and the test information vector. The plot() method for info objects can be used to visualize the IIFs and TIF (see plot.info()).

Methods (by class)

info(default): Default method to compute item and test information functions for a data frame x containing the item metadata.

info(est_item): An object created by the function est_item().

info(est_irt): An object created by the function est_irt().

References

Hambleton, R. K., & Swaminathan, H. (1985) Item response theory: Principles and applications. Boston, MA: Kluwer.

Hambleton, R. K., Swaminathan, H., & Rogers, H. J. (1991) Fundamentals of item response theory. Newbury Park, CA: Sage.

See also

Author

Hwanggyu Lim hglim83@gmail.com

Examples

## Example 1.
## Using the function `shape_df()` to create a data frame of test metadata

# Create a list of dichotomous item parameters
par.drm <- list(
  a = c(1.1, 1.2, 0.9, 1.8, 1.4),
  b = c(0.1, -1.6, -0.2, 1.0, 1.2),
  g = rep(0.2, 5)
)

# Create a list of polytomous item parameters
par.prm <- list(
  a = c(1.4, 0.6),
  d = list(c(-1.9, 0.0, 1.2), c(0.4, -1.1, 1.5, 0.2))
)

# Create a numeric vector for the number of score categories for each item
cats <- c(2, 4, 2, 2, 5, 2, 2)

# Create a character vector of IRT model types for each item
model <- c("DRM", "GRM", "DRM", "DRM", "GPCM", "DRM", "DRM")

# Create item metadata using `shape_df()`
test <- shape_df(
  par.drm = par.drm, par.prm = par.prm,
  cats = cats, model = model
) # create a data frame

# Define a sequence of theta values
theta <- seq(-2, 2, 0.1)

# Compute item and test information values based on theta
info(x = test, theta = theta, D = 1, tif = TRUE)
#> Warning: All 'DRM' items are considered as '3PLM' items during the item parameter estimation. 
#> $iif
#>         theta.1      theta.2     theta.3      theta.4     theta.5     theta.6
#> V1 0.0263743961 0.0309898876 0.036233323 0.0421430717 0.048747726 0.056062827
#> V2 0.5089696100 0.5179320384 0.523688861 0.5265585241 0.527022701 0.525689406
#> V3 0.2055811407 0.2167359208 0.226344293 0.2341622446 0.240000000 0.243731579
#> V4 0.0444475860 0.0492869567 0.054413742 0.0598029694 0.065421746 0.071229111
#> V5 0.1988148347 0.2151575434 0.232574387 0.2510619175 0.270597380 0.291134923
#> V6 0.0002562193 0.0003650065 0.000519370 0.0007379799 0.001046865 0.001482115
#> V7 0.0009317741 0.0012188932 0.001591896 0.0020752121 0.002699625 0.003503641
#>        theta.7     theta.8     theta.9    theta.10    theta.11   theta.12
#> V1 0.064087500 0.072801179 0.082160668 0.092097800 0.102517963 0.11329978
#> V2 0.523243647 0.520390856 0.517799112 0.516046021 0.515575375 0.51666745
#> V3 0.245300283 0.244719688 0.242070218 0.237491793 0.231173417 0.22334080
#> V4 0.077176168 0.083206495 0.089256893 0.095258433 0.101137833 0.10681910
#> V5 0.312601724 0.334894213 0.357874656 0.381368405 0.405162175 0.42900374
#> V6 0.002093457 0.002948877 0.004140486 0.005791734 0.008065946 0.01117590
#> V7 0.004535037 0.005852527 0.007527492 0.009645629 0.012308336 0.01563358
#>      theta.13   theta.14   theta.15   theta.16   theta.17   theta.18   theta.19
#> V1 0.12429612 0.13533673 0.14623232 0.15678031 0.16677184 0.17599979 0.18426732
#> V2 0.51942540 0.52377858 0.52950217 0.53625093 0.54360326 0.55111057 0.55834592
#> V3 0.21424316 0.20414028 0.19329078 0.18194215 0.17032295 0.15863741 0.14706208
#> V4 0.11222540 0.11728112 0.12191396 0.12605701 0.12965081 0.13264504 0.13500000
#> V5 0.45260344 0.47563788 0.49775601 0.51858777 0.53775511 0.55488514 0.56962468
#> V6 0.01539370 0.02105975 0.02858851 0.03846832 0.05125113 0.06752801 0.08788632
#> V7 0.01975592 0.02482524 0.03100377 0.03846097 0.04736580 0.05787609 0.07012521
#>      theta.20  theta.21  theta.22  theta.23   theta.24   theta.25   theta.26
#> V1 0.19139635 0.1972355 0.2016667 0.2046101 0.20602713 0.20592101 0.20433517
#> V2 0.56494542 0.5706368 0.5752511 0.5787152 0.58102885 0.58222793 0.58234411
#> V3 0.13574452 0.1248035 0.1143307 0.1043927 0.09503405 0.08628032 0.07814124
#> V4 0.13668775 0.1376927 0.1380120 0.1376548 0.13664240 0.13500637 0.13278770
#> V5 0.58165536 0.5907083 0.5965773 0.5991288 0.59830928 0.59414722 0.58675160
#> V6 0.11284591 0.1427749 0.1777903 0.2176550 0.26168698 0.30870467 0.35702757
#> V7 0.08420603 0.1001534 0.1179259 0.1373895 0.15830447 0.18031814 0.20296644
#>      theta.27   theta.28  theta.29  theta.30   theta.31   theta.32   theta.33
#> V1 0.20134967 0.19707598 0.1916505 0.1852275 0.17797207 0.17005290 0.16163653
#> V2 0.58136692 0.57921672 0.5757339 0.5706868 0.56379617 0.55477326 0.54336285
#> V3 0.07061364 0.06368422 0.0573320 0.0515304 0.04624911 0.04145546 0.03711572
#> V4 0.13003505 0.12680300 0.1231504 0.1191383 0.11482891 0.11028338 0.10556103
#> V5 0.57630568 0.56305727 0.5473062 0.5293903 0.50967033 0.48851571 0.46629162
#> V6 0.40455133 0.44890161 0.4876558 0.5186033 0.54000000 0.55077091 0.55061930
#> V7 0.22568505 0.24783151 0.2687176 0.2876501 0.30397612 0.31712853 0.32666667
#>      theta.34   theta.35   theta.36   theta.37   theta.38   theta.39   theta.40
#> V1 0.15288194 0.14393655 0.13493328 0.12598874 0.11720235 0.10865636 0.10041647
#> V2 0.52938481 0.51276673 0.49356298 0.47195792 0.44825407 0.42284839 0.39620085
#> V3 0.03319594 0.02966271 0.02648371 0.02362803 0.02106646 0.01877163 0.01671810
#> V4 0.10071804 0.09580666 0.09087461 0.08596466 0.08111441 0.07635629 0.07171764
#> V5 0.44334772 0.42000936 0.39657111 0.37329271 0.35039702 0.32806989 0.30646133
#> V6 0.54002548 0.52014019 0.49260144 0.45931563 0.42224412 0.38322664 0.34385864
#> V7 0.33230674 0.33393896 0.33162942 0.32560701 0.31623796 0.30399164 0.28940229
#>      theta.41
#> V1 0.09253296
#> V2 0.36880058
#> V3 0.01488232
#> V4 0.06722095
#> V5 0.28568780
#> V6 0.30542516
#> V7 0.27303100
#> 
#> $tif
#>  [1] 0.9853756 1.0316862 1.0753659 1.1165419 1.1555360 1.1928336 1.2290378
#>  [8] 1.2648138 1.3008295 1.3376998 1.3759410 1.4159403 1.4579432 1.5020596
#> [15] 1.5482875 1.5965475 1.6467209 1.6986821 1.7523115 1.8074813 1.8640052
#> [22] 1.9215539 1.9795462 2.0370332 2.0926057 2.1443538 2.1899073 2.2265703
#> [29] 2.2515465 2.2622268 2.2564927 2.2329802 2.1912537 2.1318607 2.0562612
#> [36] 1.9666566 1.8657547 1.7565164 1.6419208 1.5247753 1.4075808
#> 
#> $theta
#>  [1] -2.0 -1.9 -1.8 -1.7 -1.6 -1.5 -1.4 -1.3 -1.2 -1.1 -1.0 -0.9 -0.8 -0.7 -0.6
#> [16] -0.5 -0.4 -0.3 -0.2 -0.1  0.0  0.1  0.2  0.3  0.4  0.5  0.6  0.7  0.8  0.9
#> [31]  1.0  1.1  1.2  1.3  1.4  1.5  1.6  1.7  1.8  1.9  2.0
#> 
#> attr(,"class")
#> [1] "info"

## Example 2.
## Using a "-prm.txt" file exported from flexMIRT

# Import a sample "-prm.txt" output file from flexMIRT
flex_prm <- system.file("extdata", "flexmirt_sample-prm.txt",
  package = "irtQ"
)

# Read item parameters and convert them into item metadata
test_flex <- bring.flexmirt(file = flex_prm, "par")$Group1$full_df

# Define a sequence of theta values
theta <- seq(-2, 2, 0.1)

# Compute item and test information values based on theta
info(x = test_flex, theta = theta, D = 1, tif = TRUE)
#> $iif
#>            theta.1      theta.2      theta.3      theta.4      theta.5
#> CMC1  5.720247e-03 6.483134e-03 7.334591e-03 8.282216e-03 9.333729e-03
#> CMC2  1.743161e-01 2.193188e-01 2.706691e-01 3.272843e-01 3.873372e-01
#> CMC3  3.565037e-02 4.007482e-02 4.487912e-02 5.006551e-02 5.563040e-02
#> CMC4  5.657820e-02 6.403286e-02 7.204166e-02 8.055661e-02 8.950976e-02
#> CMC5  4.775355e-02 5.251382e-02 5.751685e-02 6.273759e-02 6.814466e-02
#> CMC6  4.217290e-03 5.747226e-03 7.795859e-03 1.051945e-02 1.411149e-02
#> CMC7  1.461569e-02 1.675813e-02 1.915820e-02 2.183500e-02 2.480665e-02
#> CMC8  2.235493e-02 2.507489e-02 2.804221e-02 3.126466e-02 3.474764e-02
#> CMC9  1.025995e-02 1.173138e-02 1.337896e-02 1.521614e-02 1.725578e-02
#> CMC10 1.759820e-02 2.273209e-02 2.917233e-02 3.717008e-02 4.699272e-02
#> CMC11 7.814624e-02 8.640578e-02 9.502239e-02 1.039211e-01 1.130124e-01
#> CMC12 1.570051e-02 1.784647e-02 2.022905e-02 2.286355e-02 2.576431e-02
#> CMC13 4.467658e-04 5.931060e-04 7.864289e-04 1.041322e-03 1.376649e-03
#> CMC14 7.550703e-03 9.896514e-03 1.291070e-02 1.675475e-02 2.161596e-02
#> CMC15 5.873680e-02 6.966697e-02 8.202991e-02 9.585456e-02 1.111273e-01
#> CMC16 2.131068e-02 2.990318e-02 4.149633e-02 5.689010e-02 7.698302e-02
#> CMC17 2.996996e-02 3.720264e-02 4.585503e-02 5.609042e-02 6.805223e-02
#> CMC18 1.234103e-04 1.725103e-04 2.409608e-04 3.362683e-04 4.687743e-04
#> CMC19 1.326868e-01 1.836619e-01 2.482572e-01 3.269266e-01 4.185116e-01
#> CMC20 3.036616e-01 3.235275e-01 3.400048e-01 3.524906e-01 3.605602e-01
#> CMC21 2.678725e-01 3.072321e-01 3.471277e-01 3.862077e-01 4.229760e-01
#> CMC22 6.531495e-02 7.008860e-02 7.488982e-02 7.967304e-02 8.438964e-02
#> CMC23 6.311253e-02 7.016026e-02 7.761115e-02 8.542034e-02 9.353019e-02
#> CMC24 7.867192e-04 9.784501e-04 1.215470e-03 1.507928e-03 1.868042e-03
#> CMC25 7.673668e-02 7.990170e-02 8.283922e-02 8.551524e-02 8.789897e-02
#> CMC26 2.063825e-01 2.112304e-01 2.145813e-01 2.163858e-01 2.166352e-01
#> CMC27 3.538070e-02 4.177404e-02 4.904638e-02 5.724673e-02 6.640844e-02
#> CMC28 4.185911e-03 6.282282e-03 9.376392e-03 1.390140e-02 2.044670e-02
#> CMC29 1.944752e-01 2.112001e-01 2.269281e-01 2.412170e-01 2.536523e-01
#> CMC30 1.667968e-03 2.151761e-03 2.770037e-03 3.557511e-03 4.556661e-03
#> CMC31 3.084686e-02 3.409533e-02 3.757680e-02 4.129144e-02 4.523671e-02
#> CMC32 9.450542e-02 1.140193e-01 1.358531e-01 1.597414e-01 1.852410e-01
#> CMC33 1.890363e-01 2.060739e-01 2.223409e-01 2.374022e-01 2.508369e-01
#> CMC34 1.949609e-02 2.406976e-02 2.955309e-02 3.607090e-02 4.374702e-02
#> CMC35 2.136860e-02 2.714797e-02 3.425525e-02 4.290010e-02 5.328877e-02
#> CMC36 1.297331e-02 1.487304e-02 1.700361e-02 1.938304e-02 2.202874e-02
#> CMC37 5.736938e-02 7.396089e-02 9.428816e-02 1.187717e-01 1.477244e-01
#> CMC38 4.354857e-02 4.662214e-02 4.974907e-02 5.290894e-02 5.607939e-02
#> CFR1  9.832480e-01 1.028761e+00 1.066003e+00 1.095460e+00 1.118084e+00
#> CFR2  2.257208e-01 2.460116e-01 2.669861e-01 2.884416e-01 3.101421e-01
#> AMC1  2.818814e-03 3.705104e-03 4.856111e-03 6.344043e-03 8.257555e-03
#> AMC2  4.040393e-01 4.494884e-01 4.903920e-01 5.245960e-01 5.502632e-01
#> AMC3  6.780115e-03 8.767172e-03 1.128935e-02 1.446997e-02 1.845221e-02
#> AMC4  7.179654e-02 7.936298e-02 8.728349e-02 9.549832e-02 1.039349e-01
#> AMC5  7.735752e-04 9.335349e-04 1.125437e-03 1.355298e-03 1.630157e-03
#> AMC6  1.657259e-06 2.609888e-06 4.109573e-06 6.469944e-06 1.018392e-05
#> AMC7  4.282076e-02 5.074714e-02 5.976158e-02 6.991249e-02 8.122370e-02
#> AMC8  8.093630e-03 1.086171e-02 1.450016e-02 1.924278e-02 2.536654e-02
#> AMC9  3.543831e-02 4.198605e-02 4.946823e-02 5.794896e-02 6.748006e-02
#> AMC10 2.532889e-03 3.226568e-03 4.099219e-03 5.192523e-03 6.556125e-03
#> AMC11 1.199164e-01 1.487875e-01 1.816598e-01 2.180225e-01 2.569707e-01
#> AMC12 7.247179e-02 7.956084e-02 8.683951e-02 9.422520e-02 1.016246e-01
#> AFR1  1.529539e-01 1.661563e-01 1.799645e-01 1.943055e-01 2.090876e-01
#> AFR2  4.453958e-01 4.545051e-01 4.622439e-01 4.686948e-01 4.739551e-01
#> AFR3  1.482069e-01 1.552201e-01 1.621768e-01 1.690334e-01 1.757465e-01
#>            theta.6      theta.7      theta.8      theta.9     theta.10
#> CMC1  1.049685e-02 1.177918e-02 1.318800e-02 1.473014e-02 1.641172e-02
#> CMC2  4.482809e-01 5.069979e-01 5.600758e-01 6.041834e-01 6.364840e-01
#> CMC3  6.156357e-02 6.784740e-02 7.445624e-02 8.135591e-02 8.850335e-02
#> CMC4  9.881252e-02 1.083561e-01 1.180127e-01 1.276384e-01 1.370761e-01
#> CMC5  7.370030e-02 7.936055e-02 8.507565e-02 9.079064e-02 9.644620e-02
#> CMC6  1.880697e-02 2.488511e-02 3.266942e-02 4.252407e-02 5.484523e-02
#> CMC7  2.808972e-02 3.169861e-02 3.564481e-02 3.993624e-02 4.457643e-02
#> CMC8  3.849373e-02 4.250222e-02 4.676864e-02 5.128438e-02 5.603632e-02
#> CMC9  1.950971e-02 2.198827e-02 2.469979e-02 2.765005e-02 3.084168e-02
#> CMC10 5.891221e-02 7.318837e-02 9.004718e-02 1.096544e-01 1.320861e-01
#> CMC11 1.221929e-01 1.313469e-01 1.403489e-01 1.490664e-01 1.573632e-01
#> CMC12 2.894428e-02 3.241457e-02 3.618394e-02 4.025830e-02 4.464016e-02
#> CMC13 1.816673e-03 2.392417e-03 3.143266e-03 4.118832e-03 5.381029e-03
#> CMC14 2.770562e-02 3.525412e-02 4.450182e-02 5.568497e-02 6.901607e-02
#> CMC15 1.277829e-01 1.456973e-01 1.646812e-01 1.844773e-01 2.047607e-01
#> CMC16 1.027363e-01 1.351167e-01 1.750150e-01 2.231395e-01 2.798839e-01
#> CMC17 8.184913e-02 9.753835e-02 1.151081e-01 1.344605e-01 1.553977e-01
#> CMC18 6.526772e-04 9.073908e-04 1.259333e-03 1.744246e-03 2.410164e-03
#> CMC19 5.197740e-01 6.252285e-01 7.274532e-01 8.179557e-01 8.884983e-01
#> CMC20 3.639975e-01 3.628050e-01 3.571945e-01 3.475597e-01 3.344350e-01
#> CMC21 4.558944e-01 4.835052e-01 5.045600e-01 5.181361e-01 5.237221e-01
#> CMC22 8.898888e-02 9.341887e-02 9.762771e-02 1.015647e-01 1.051817e-01
#> CMC23 1.018703e-01 1.103577e-01 1.188982e-01 1.273873e-01 1.357128e-01
#> CMC24 2.310445e-03 2.852561e-03 3.515003e-03 4.321978e-03 5.301681e-03
#> CMC25 8.996352e-02 9.168653e-02 9.305070e-02 9.404411e-02 9.466040e-02
#> CMC26 2.153615e-01 2.126341e-01 2.085552e-01 2.032543e-01 1.968811e-01
#> CMC27 7.654456e-02 8.764298e-02 9.966196e-02 1.125261e-01 1.261235e-01
#> CMC28 2.979130e-02 4.292777e-02 6.106322e-02 8.557794e-02 1.179204e-01
#> CMC29 2.638719e-01 2.715876e-01 2.766017e-01 2.788178e-01 2.782436e-01
#> CMC30 5.818952e-03 7.406007e-03 9.390615e-03 1.185740e-02 1.490289e-02
#> CMC31 4.940696e-02 5.379316e-02 5.838266e-02 6.315891e-02 6.810131e-02
#> CMC32 2.117264e-01 2.384059e-01 2.643589e-01 2.885934e-01 3.101200e-01
#> CMC33 2.622620e-01 2.713546e-01 2.778709e-01 2.816596e-01 2.826690e-01
#> CMC34 5.269770e-02 6.302355e-02 7.479997e-02 8.806668e-02 1.028168e-01
#> CMC35 6.560844e-02 8.000734e-02 9.657154e-02 1.152999e-01 1.360794e-01
#> CMC36 2.495701e-02 2.818249e-02 3.171759e-02 3.557180e-02 3.975103e-02
#> CMC37 1.812901e-01 2.193777e-01 2.615989e-01 3.072178e-01 3.551232e-01
#> CMC38 5.923640e-02 6.235459e-02 6.540760e-02 6.836850e-02 7.121027e-02
#> CFR1  1.135045e+00 1.147499e+00 1.156424e+00 1.162541e+00 1.166286e+00
#> CFR2  3.318248e-01 3.532093e-01 3.740098e-01 3.939484e-01 4.127702e-01
#> AMC1  1.070399e-02 1.381122e-02 1.772872e-02 2.262722e-02 2.869650e-02
#> AMC2  5.660749e-01 5.713720e-01 5.662105e-01 5.513210e-01 5.279867e-01
#> AMC3  2.339875e-02 2.949007e-02 3.692041e-02 4.589140e-02 5.660256e-02
#> AMC4  1.125082e-01 1.211218e-01 1.296693e-01 1.380366e-01 1.461045e-01
#> AMC5  1.958211e-03 2.348966e-03 2.813387e-03 3.364054e-03 4.015310e-03
#> AMC6  1.602574e-05 2.521046e-05 3.964314e-05 6.230681e-05 9.786511e-05
#> AMC7  9.368794e-02 1.072606e-01 1.218542e-01 1.373340e-01 1.535156e-01
#> AMC8  3.319017e-02 4.306768e-02 5.537484e-02 7.048702e-02 8.874729e-02
#> AMC9  7.809656e-02 8.981192e-02 1.026132e-01 1.164561e-01 1.312608e-01
#> AMC10 8.248525e-03 1.033783e-02 1.290227e-02 1.603041e-02 1.982082e-02
#> AMC11 2.972018e-01 3.370673e-01 3.746874e-01 4.081175e-01 4.355472e-01
#> AMC12 1.089356e-01 1.160498e-01 1.228557e-01 1.292418e-01 1.351006e-01
#> AFR1  2.242007e-01 2.395173e-01 2.548958e-01 2.701831e-01 2.852203e-01
#> AFR2  4.781256e-01 4.813008e-01 4.835614e-01 4.849670e-01 4.855525e-01
#> AFR3  1.822732e-01 1.885729e-01 1.946076e-01 2.003432e-01 2.057496e-01
#>           theta.11     theta.12     theta.13     theta.14     theta.15
#> CMC1  0.0182379577 0.0202128951 0.0223391692 0.0246177332 0.0270476000
#> CMC2  0.6550073917 0.6588985800 0.6484896965 0.6251866852 0.5912069926
#> CMC3  0.0958465026 0.1033245124 0.1108681273 0.1184005066 0.1258383456
#> CMC4  0.1461608405 0.1547247706 0.1626034894 0.1696422709 0.1757022636
#> CMC5  0.1019797363 0.1073265921 0.1124215004 0.1172001159 0.1216006248
#> CMC6  0.0700463858 0.0885368114 0.1106926347 0.1368206688 0.1671157253
#> CMC7  0.0495638000 0.0548908925 0.0605437055 0.0665010791 0.0727342164
#> CMC8  0.0610064352 0.0661716452 0.0715036229 0.0769687907 0.0825284348
#> CMC9  0.0342736036 0.0379404746 0.0418321588 0.0459333130 0.0502230604
#> CMC10 0.1572969854 0.1850918433 0.2151010523 0.2467666818 0.2793425039
#> CMC11 0.1651038756 0.1721575853 0.1784028448 0.1837316366 0.1880533255
#> CMC12 0.0493280998 0.0543162094 0.0595936142 0.0651439931 0.0709451869
#> CMC13 0.0070063072 0.0090879234 0.0117380288 0.0150892956 0.0192956734
#> CMC14 0.0846589093 0.1026991918 0.1231131472 0.1457375315 0.1702456045
#> CMC15 0.2251440228 0.2451873857 0.2644135726 0.2823280683 0.2984428275
#> CMC16 0.3451741417 0.4183037335 0.4977799838 0.5812141184 0.6653011771
#> CMC17 0.1776117513 0.2006827960 0.2240850833 0.2472038638 0.2693624373
#> CMC18 0.0033211257 0.0045616949 0.0062422936 0.0085051876 0.0115307621
#> CMC19 0.9325797527 0.9466543537 0.9307104424 0.8880471010 0.8243747242
#> CMC20 0.3184480180 0.3002697685 0.2805700616 0.2599802009 0.2390658439
#> CMC21 0.5212596225 0.5111339743 0.4941173080 0.4712755180 0.4438553250
#> CMC22 0.1084339716 0.1112817358 0.1136909174 0.1156340809 0.1170910687
#> CMC23 0.1437566772 0.1513988947 0.1585202188 0.1650060974 0.1707503056
#> CMC24 0.0064866426 0.0079140030 0.0096256495 0.0116681677 0.0140925327
#> CMC25 0.0948988342 0.0947641028 0.0942660818 0.0934193898 0.0922428633
#> CMC26 0.1895987519 0.1815768550 0.1729850412 0.1639875159 0.1547385385
#> CMC27 0.1403037434 0.1548785071 0.1696233791 0.1842825423 0.1985756985
#> CMC28 0.1594193381 0.2110088638 0.2728842312 0.3441442764 0.4225091538
#> CMC29 0.2749858066 0.2692394274 0.2612709447 0.2513987410 0.2399722596
#> CMC30 0.0186347452 0.0231697461 0.0286302327 0.0351386475 0.0428099443
#> CMC31 0.0731850882 0.0783812959 0.0836568525 0.0889747319 0.0942942543
#> CMC32 0.3280326945 0.3415868637 0.3502629527 0.3538067917 0.3522414487
#> CMC33 0.2809470421 0.2766340817 0.2699502392 0.2611781700 0.2506436197
#> CMC34 0.1189860782 0.1364440819 0.1549871426 0.1743355900 0.1941355756
#> CMC35 0.1586647553 0.1826649584 0.2075414559 0.2326204300 0.2571209270
#> CMC36 0.0442569304 0.0490861746 0.0542298109 0.0596726622 0.0653928126
#> CMC37 0.4038361000 0.4515612492 0.4962890822 0.5359441942 0.5685678539
#> CMC38 0.0739062819 0.0764307813 0.0787594193 0.0808697149 0.0827415078
#> CFR1  1.1678232555 1.1670415109 1.1635587673 1.1567240495 1.1456441369
#> CFR2  0.4302559515 0.4462336961 0.4605857940 0.4732516878 0.4842255290
#> AMC1  0.0361403546 0.0451682840 0.0559830915 0.0687643076 0.0836475665
#> AMC2  0.4978671036 0.4628013548 0.4246263573 0.3850310484 0.3454592830
#> AMC3  0.0692386888 0.0839541049 0.1008542537 0.1199755036 0.1412645936
#> AMC4  0.1537516379 0.1608581678 0.1673093083 0.1729991144 0.1778340328
#> AMC5  0.0047833986 0.0056865815 0.0067452151 0.0079817856 0.0094208757
#> AMC6  0.0001535947 0.0002408212 0.0003771174 0.0005896437 0.0009201758
#> AMC7  0.1701649368 0.1870013933 0.2037037513 0.2199200637 0.2352805345
#> AMC8  0.1104248218 0.1356654299 0.1644380057 0.1964831594 0.2312722017
#> AMC9  0.1469081705 0.1632372428 0.1800444755 0.1970848503 0.2140755160
#> AMC10 0.0243811657 0.0298265626 0.0362769922 0.0438537684 0.0526749269
#> AMC11 0.4554991609 0.4669923700 0.4696393402 0.4636620067 0.4498273849
#> AMC12 0.1403318947 0.1448466178 0.1485697542 0.1514430194 0.1534267172
#> AFR1  0.2998472110 0.3139090658 0.3272623494 0.3397805321 0.3513588135
#> AFR2  0.4853253475 0.4842649188 0.4823240371 0.4794321491 0.4755004590
#> AFR3  0.2108017985 0.2154798703 0.2197690543 0.2236596200 0.2271464462
#>          theta.16    theta.17    theta.18    theta.19    theta.20   theta.21
#> CMC1  0.029625593 0.032346115 0.035200955 0.038179131 0.041266790 0.04444717
#> CMC2  0.549236771 0.502083190 0.452382867 0.402400573 0.353924743 0.30824605
#> CMC3  0.133093351 0.140074044 0.146687856 0.152843446 0.158453164 0.16343555
#> CMC4  0.180666133 0.184442776 0.186970793 0.188220476 0.188194218 0.18692535
#> CMC5  0.125565355 0.129042325 0.131986652 0.134361770 0.136140386 0.13730514
#> CMC6  0.201613112 0.240139245 0.282264906 0.327267507 0.374110460 0.42144875
#> CMC7  0.079206360 0.085872665 0.092680306 0.099568843 0.106470883 0.11331303
#> CMC8  0.088138974 0.093752392 0.099316840 0.104777409 0.110077059 0.11515770
#> CMC9  0.054674808 0.059256230 0.063929447 0.068651412 0.073374507 0.07804737
#> CMC10 0.311911942 0.343426109 0.372761659 0.398795014 0.420486139 0.43696228
#> CMC11 0.191297919 0.193418498 0.194392672 0.194222969 0.192936152 0.19058154
#> CMC12 0.076968894 0.083180484 0.089538950 0.095997039 0.102501558 0.10899390
#> CMC13 0.024531771 0.030990270 0.038876730 0.048401184 0.059766064 0.07315027
#> CMC14 0.196134215 0.222726862 0.249196272 0.274607602 0.297980080 0.31836137
#> CMC15 0.312302239 0.323509138 0.331748453 0.336806106 0.338581176 0.33709008
#> CMC16 0.745937623 0.818510698 0.878358098 0.921344041 0.944444948 0.94620931
#> CMC17 0.289858475 0.308006950 0.323185724 0.334878977 0.342713477 0.34648340
#> CMC18 0.015543393 0.020815780 0.027670090 0.036473702 0.047626950 0.06154014
#> CMC19 0.746575509 0.661504097 0.575100972 0.491920242 0.415028748 0.34615855
#> CMC20 0.218310106 0.198106031 0.178756717 0.160481079 0.143423235 0.12766378
#> CMC21 0.413169275 0.380493639 0.346989178 0.313648979 0.281272612 0.25046244
#> CMC22 0.118049424 0.118504575 0.118459767 0.117925772 0.116920370 0.11546765
#> CMC23 0.175658502 0.179651458 0.182667762 0.184665857 0.185625266 0.18554696
#> CMC24 0.016953462 0.020308355 0.024215737 0.028733154 0.033914476 0.03980663
#> CMC25 0.090758944 0.088993009 0.086972680 0.084727121 0.082286370 0.07968070
#> CMC26 0.145379029 0.136034277 0.126812663 0.117805271 0.109086226 0.10071359
#> CMC27 0.212207333 0.224877950 0.236296711 0.246194663 0.254337626 0.26053771
#> CMC28 0.504224267 0.584249423 0.656776469 0.716023385 0.757146853 0.77704579
#> CMC29 0.227351940 0.213891431 0.199923095 0.185747297 0.171625510 0.15777690
#> CMC30 0.051741808 0.062002896 0.073619690 0.086562941 0.100735147 0.11596081
#> CMC31 0.099571505 0.104759883 0.109810772 0.114674335 0.119300406 0.12363947
#> CMC32 0.345850201 0.335135017 0.320758374 0.303477776 0.284081907 0.26333540
#> CMC33 0.238695508 0.225687256 0.211960643 0.197832994 0.183588017 0.16947022
#> CMC34 0.213966609 0.233355195 0.251794383 0.268768341 0.283780298 0.29638156
#> CMC35 0.280198299 0.300999928 0.318727781 0.332700541 0.342407426 0.34754661
#> CMC36 0.071361218 0.077541475 0.083889775 0.090355079 0.096879542 0.10339920
#> CMC37 0.592510777 0.606605905 0.610290737 0.603656627 0.587416741 0.56280102
#> CMC38 0.084357358 0.085702882 0.086767021 0.087542221 0.088024526 0.08821360
#> CFR1  1.129247627 1.106398235 1.076058705 1.037488438 0.990435885 0.93527374
#> CFR2  0.493549178 0.501301500 0.507585304 0.512513383 0.516195032 0.51872414
#> AMC1  0.100700889 0.119899614 0.141102582 0.164032925 0.188267173 0.21323629
#> AMC2  0.307061109 0.270684852 0.236898617 0.206029445 0.178210110 0.15342620
#> AMC3  0.164559621 0.189574906 0.215892333 0.242961753 0.270112619 0.29657817
#> AMC4  0.181736114 0.184645683 0.186523308 0.187350953 0.187132240 0.18589184
#> AMC5  0.011089050 0.013014635 0.015227387 0.017758005 0.020637502 0.02389640
#> AMC6  0.001432581 0.002223763 0.003439378 0.005295799 0.008109804 0.01233683
#> AMC7  0.249413145 0.261961073 0.272600663 0.281058517 0.287126193 0.29067117
#> AMC8  0.267985836 0.305521543 0.342536247 0.377526230 0.408939816 0.43531122
#> AMC9  0.230702140 0.246628072 0.261506166 0.274992775 0.286763164 0.29652727
#> AMC10 0.062849533 0.074470940 0.087609105 0.102302152 0.118547497 0.13629294
#> AMC11 0.429321310 0.403588434 0.374168551 0.342553948 0.310082956 0.27787470
#> AMC12 0.154500807 0.154665116 0.153938708 0.152358482 0.149977113 0.14686050
#> AFR1  0.361917463 0.371403413 0.379789929 0.387074394 0.393274418 0.39842267
#> AFR2  0.470428867 0.464114503 0.456461453 0.447391009 0.436851618 0.42482754
#> AFR3  0.230228417 0.232907661 0.235188678 0.237077419 0.238580370 0.23970370
#>         theta.22   theta.23   theta.24   theta.25   theta.26   theta.27
#> CMC1  0.04770066 0.05100487 0.05433490 0.05766359 0.06096192 0.06419948
#> CMC2  0.26619517 0.22821462 0.19444358 0.16480102 0.13905864 0.11689991
#> CMC3  0.16771778 0.17123787 0.17394661 0.17580911 0.17680574 0.17693263
#> CMC4  0.18447557 0.18093115 0.17639832 0.17099812 0.16486109 0.15812216
#> CMC5  0.13784894 0.13777494 0.13709622 0.13583513 0.13402239 0.13169592
#> CMC6  0.46766917 0.51097062 0.54948402 0.58142319 0.60524938 0.61982551
#> CMC7  0.12001714 0.12650187 0.13268449 0.13848283 0.14381743 0.14861374
#> CMC8  0.11996138 0.12443158 0.12851449 0.13216038 0.13532479 0.13796972
#> CMC9  0.08261590 0.08702445 0.09121719 0.09513947 0.09873928 0.10196869
#> CMC10 0.44759072 0.45203029 0.45025431 0.44254203 0.42944101 0.41170696
#> CMC11 0.18722844 0.18296296 0.17788426 0.17210070 0.16572590 0.15887506
#> CMC12 0.11541079 0.12168526 0.12774779 0.13352774 0.13895483 0.14396085
#> CMC13 0.08868967 0.10645480 0.12642741 0.14847791 0.17234675 0.19763278
#> CMC14 0.33490587 0.34694659 0.35405044 0.35604930 0.35304313 0.34537630
#> CMC15 0.33246239 0.32492901 0.31480411 0.30246302 0.28831830 0.27279638
#> CMC16 0.92697221 0.88876677 0.83496028 0.76971613 0.69741570 0.62215993
#> CMC17 0.34616083 0.34189098 0.33397352 0.32283276 0.30898107 0.29297999
#> CMC18 0.07859767 0.09910857 0.12324510 0.15097475 0.18199474 0.21568166
#> CMC19 0.28598550 0.23443518 0.19095697 0.15474032 0.12486991 0.10042734
#> CMC20 0.11323157 0.10011510 0.08827269 0.07764145 0.06814460 0.05969738
#> CMC21 0.22163621 0.19504995 0.17082578 0.14898075 0.12945378 0.11212932
#> CMC22 0.11359718 0.11134303 0.10874279 0.10583651 0.10266573 0.09927247
#> CMC23 0.18445284 0.18238440 0.17940065 0.17557551 0.17099473 0.16575263
#> CMC24 0.04644578 0.05385317 0.06203069 0.07095658 0.08058158 0.09082581
#> CMC25 0.07694006 0.07409354 0.07116900 0.06819268 0.06518892 0.06218003
#> CMC26 0.09273065 0.08516747 0.07804256 0.07136454 0.06513385 0.05934421
#> CMC27 0.26466251 0.26664119 0.26646691 0.26419556 0.25994083 0.25386633
#> CMC28 0.77483727 0.75188214 0.71138143 0.65769208 0.59556669 0.52949892
#> CMC29 0.14437786 0.13156380 0.11943255 0.10804873 0.09744861 0.08764504
#> CMC30 0.13198137 0.14845661 0.16497376 0.18106483 0.19623139 0.20997505
#> CMC31 0.12764371 0.13126804 0.13447114 0.13721649 0.13947320 0.14121682
#> CMC32 0.24193648 0.22048905 0.19948842 0.17931849 0.16025761 0.14248983
#> CMC33 0.15568250 0.14238641 0.12970433 0.11772310 0.10649839 0.09605944
#> CMC34 0.30619892 0.31295760 0.31649752 0.31678092 0.31389097 0.30802161
#> CMC35 0.34804335 0.34404607 0.33590212 0.32411766 0.30930817 0.29214623
#> CMC36 0.10984489 0.11614354 0.12221954 0.12799649 0.13339899 0.13835461
#> CMC37 0.53140044 0.49499013 0.45536039 0.41417761 0.37288717 0.33266030
#> CMC38 0.08811262 0.08772815 0.08706989 0.08615037 0.08498464 0.08358985
#> CFR1  0.87303228 0.80531430 0.73411129 0.66156779 0.58974895 0.52045526
#> CFR2  0.52016962 0.52056865 0.51992305 0.51819889 0.51532939 0.51122102
#> AMC1  0.23824138 0.26248519 0.28511839 0.30529711 0.32224573 0.33531763
#> AMC2  0.13155891 0.11242112 0.09578618 0.08140961 0.06904471 0.05845311
#> AMC3  0.32153212 0.34413600 0.36359358 0.37920694 0.39042786 0.39689832
#> AMC4  0.18367402 0.18054047 0.17656761 0.17184343 0.16646418 0.16053109
#> AMC5  0.02756375 0.03166598 0.03622563 0.04125982 0.04677883 0.05278443
#> AMC6  0.01861710 0.02782564 0.04111711 0.05994864 0.08605570 0.12134878
#> AMC7  0.29164303 0.29007441 0.28607660 0.27983046 0.27157372 0.26158587
#> AMC8  0.45539784 0.46830039 0.47354631 0.47112347 0.46145994 0.44535601
#> AMC9  0.30404458 0.30913686 0.31169751 0.31169674 0.30918203 0.30427386
#> AMC10 0.15542830 0.17577836 0.19709788 0.21906979 0.24130744 0.26336172
#> AMC11 0.24680200 0.21749436 0.19036136 0.16562728 0.14336959 0.12355618
#> AMC12 0.14308489 0.13873383 0.13389518 0.12865829 0.12311139 0.11733944
#> AFR1  0.40256094 0.40573394 0.40798343 0.40934311 0.40983471 0.40946551
#> AFR2  0.41134525 0.39647685 0.38033987 0.36309370 0.34493266 0.32607684
#> AFR3  0.24045252 0.24083031 0.24083852 0.24047638 0.23974097 0.23862747
#>         theta.28   theta.29   theta.30   theta.31   theta.32   theta.33
#> CMC1  0.06734497 0.07036685 0.07323390 0.07591592 0.07838437 0.08061300
#> CMC2  0.09796491 0.08188225 0.06829049 0.05685132 0.04725676 0.03923212
#> CMC3  0.17620154 0.17463915 0.17228592 0.16919439 0.16542728 0.16105523
#> CMC4  0.15091592 0.14337258 0.13561463 0.12775436 0.11989212 0.11211536
#> CMC5  0.12889962 0.12568196 0.12209457 0.11819089 0.11402484 0.10964970
#> CMC6  0.62453413 0.61933787 0.60477116 0.58186591 0.55202619 0.51687546
#> CMC7  0.15280418 0.15633019 0.15914390 0.16120956 0.16250452 0.16301976
#> CMC8  0.14006455 0.14158688 0.14252298 0.14286812 0.14262650 0.14181104
#> CMC9  0.10478515 0.10715264 0.10904264 0.11043482 0.11131748 0.11168762
#> CMC10 0.39023099 0.36596420 0.33984899 0.31276316 0.28548037 0.25864701
#> CMC11 0.15166160 0.14419433 0.13657510 0.12889710 0.12124363 0.11368750
#> CMC12 0.14848130 0.15245712 0.15583629 0.15857528 0.16064027 0.16200808
#> CMC13 0.22379192 0.25014868 0.27592218 0.30026633 0.32232187 0.34127569
#> CMC14 0.33359205 0.31837342 0.30047951 0.28068514 0.25973017 0.23828173
#> CMC15 0.25631646 0.23927320 0.22202359 0.20487851 0.18809833 0.17189216
#> CMC16 0.54742363 0.47588150 0.40938262 0.34902871 0.29530867 0.24825040
#> CMC17 0.27540313 0.25680423 0.23769221 0.21851405 0.19964526 0.18138690
#> CMC18 0.25107056 0.28687642 0.32156539 0.35347457 0.38096771 0.40260437
#> CMC19 0.08055155 0.06446990 0.05150977 0.04109822 0.03275509 0.02608280
#> CMC20 0.05221164 0.04559928 0.03977467 0.03465637 0.03016815 0.02623958
#> CMC21 0.09685692 0.08346673 0.07178124 0.06162376 0.05282426 0.04522308
#> CMC22 0.09569843 0.09198414 0.08816838 0.08428760 0.08037555 0.07646299
#> CMC23 0.15994886 0.15368515 0.14706250 0.14017858 0.13312572 0.12598919
#> CMC24 0.10157705 0.11269049 0.12399055 0.13527489 0.14632045 0.15689166
#> CMC25 0.05918614 0.05622511 0.05331260 0.05046205 0.04768481 0.04499025
#> CMC26 0.05398410 0.04903795 0.04448726 0.04031151 0.03648892 0.03299711
#> CMC27 0.24617540 0.23709967 0.22688725 0.21579141 0.20406053 0.19192966
#> CMC28 0.46327957 0.39978463 0.34095595 0.28790837 0.24109838 0.20050507
#> CMC29 0.07863209 0.07038939 0.06288588 0.05608302 0.04993747 0.04440323
#> CMC30 0.22183071 0.23139894 0.23837374 0.24256257 0.24389631 0.24242856
#> CMC31 0.14242989 0.14310238 0.14323185 0.14282344 0.14188964 0.14044984
#> CMC32 0.12611933 0.11118560 0.09767814 0.08554971 0.07472765 0.06512317
#> CMC33 0.08641379 0.07755170 0.06945025 0.06207680 0.05539203 0.04935243
#> CMC34 0.29946019 0.28856505 0.27574057 0.26141225 0.24600391 0.22991856
#> CMC35 0.27331289 0.25345675 0.23316353 0.21293663 0.19318781 0.17423631
#> CMC36 0.14279578 0.14666173 0.14990021 0.15246896 0.15433692 0.15548501
#> CMC37 0.29437963 0.25865390 0.22585126 0.19614189 0.16954257 0.14595832
#> CMC38 0.08198484 0.08018977 0.07822566 0.07611401 0.07387646 0.07153437
#> CFR1  0.45510758 0.39470386 0.33983425 0.29073487 0.24736056 0.20946121
#> CFR2  0.50576261 0.49883685 0.49033359 0.48016364 0.46827199 0.45464898
#> AMC1  0.34404593 0.34817773 0.34768743 0.34276831 0.33380449 0.32132845
#> AMC2  0.04941143 0.04171497 0.03517939 0.02964080 0.02495510 0.02099661
#> AMC3  0.39847479 0.39523349 0.38745623 0.37559957 0.36025193 0.34208457
#> AMC4  0.15414712 0.14741412 0.14043031 0.13328821 0.12607306 0.11886170
#> AMC5  0.05926837 0.06621093 0.07357966 0.08132843 0.08939684 0.09771017
#> AMC6  0.16769963 0.22659843 0.29869506 0.38328582 0.47786310 0.57788785
#> AMC7  0.25017212 0.23764776 0.22432405 0.21049639 0.19643519 0.18237968
#> AMC8  0.42388264 0.39826474 0.36976724 0.33959807 0.30883653 0.27838935
#> AMC9  0.29715729 0.28807013 0.27728903 0.26511462 0.25185688 0.23782193
#> AMC10 0.28473384 0.30489364 0.32330294 0.33944281 0.35284257 0.36310801
#> AMC11 0.10607840 0.09077845 0.07747093 0.06595895 0.05604558 0.04754158
#> AMC12 0.11142234 0.10543358 0.09943930 0.09349776 0.08765910 0.08196545
#> AFR1  0.40822745 0.40609803 0.40304274 0.39901900 0.39398125 0.38788683
#> AFR2  0.30676141 0.28722564 0.26770254 0.24840990 0.22954332 0.21127133
#> AFR3  0.23712962 0.23524035 0.23295259 0.23026014 0.22715857 0.22364613
#>         theta.34   theta.35   theta.36   theta.37    theta.38    theta.39
#> CMC1  0.08257841 0.08426056 0.08564323 0.08671430 0.087465939 0.087894763
#> CMC2  0.03253622 0.02695984 0.02232322 0.01847309 0.015279510 0.012632914
#> CMC3  0.15615453 0.15080483 0.14508690 0.13908068 0.132863457 0.126508467
#> CMC4  0.10449826 0.09710199 0.08997524 0.08315519 0.076668594 0.070533001
#> CMC5  0.10511702 0.10047580 0.09577177 0.09104692 0.086339143 0.081682030
#> CMC6  0.47810184 0.43732271 0.39598269 0.35529009 0.316189811 0.279365781
#> CMC7  0.16275998 0.16174310 0.15999946 0.15757056 0.154507412 0.150868833
#> CMC8  0.14044276 0.13855011 0.13616796 0.13333657 0.130100399 0.126506918
#> CMC9  0.11155084 0.11092087 0.10981887 0.10827258 0.106315277 0.103984645
#> CMC10 0.23277399 0.20824003 0.18530327 0.16411755 0.144750895 0.127203839
#> CMC11 0.10629078 0.09910496 0.09217147 0.08552232 0.079180948 0.073163186
#> CMC12 0.16266678 0.16261589 0.16186628 0.16043960 0.158367501 0.155690488
#> CMC13 0.35642116 0.36721216 0.37330401 0.37457603 0.371133065 0.363286775
#> CMC14 0.21691033 0.19607882 0.17614173 0.15735231 0.139874346 0.123796400
#> CMC15 0.15641987 0.14179610 0.12809549 0.11535858 0.103597744 0.092802879
#> CMC16 0.20756403 0.17276282 0.14325680 0.11842051 0.097638385 0.080332796
#> CMC17 0.16396775 0.14755013 0.13223804 0.11808627 0.105109677 0.093292035
#> CMC18 0.41729444 0.42441048 0.42383873 0.41596245 0.401585267 0.381813239
#> CMC19 0.02075518 0.01650662 0.01312195 0.01042764 0.008284245 0.006579968
#> CMC20 0.02280628 0.01980990 0.01719792 0.01492335 0.012944355 0.011223847
#> CMC21 0.03867312 0.03304076 0.02820605 0.02406229 0.020515311 0.017482479
#> CMC22 0.07257747 0.06874332 0.06498164 0.06131032 0.057744284 0.054295569
#> CMC23 0.11884613 0.11176478 0.10480415 0.09801407 0.091435422 0.085100688
#> CMC24 0.16675017 0.17566551 0.18342610 0.18984953 0.194791413 0.198152081
#> CMC25 0.04238585 0.03987740 0.03746914 0.03516391 0.032963321 0.030867903
#> CMC26 0.02981357 0.02691610 0.02428306 0.02189366 0.019728056 0.017767483
#> CMC27 0.17961412 0.16730503 0.15516681 0.14333627 0.131923119 0.121011482
#> CMC28 0.16579325 0.13644516 0.11185819 0.09141183 0.074510214 0.060606170
#> CMC29 0.03943335 0.03498122 0.03100151 0.02745087 0.024288335 0.021475653
#> CMC30 0.23832489 0.23184406 0.22331401 0.21310559 0.201607085 0.189201487
#> CMC31 0.13852971 0.13616044 0.13337786 0.13022147 0.126733463 0.122957739
#> CMC32 0.05663872 0.04917347 0.04262747 0.03690440 0.031913471 0.027570481
#> CMC33 0.04391226 0.03902510 0.03464509 0.03072778 0.027230696 0.024113822
#> CMC34 0.21352375 0.19714179 0.18104444 0.16545146 0.150532224 0.136409614
#> CMC35 0.15631424 0.13957592 0.12410953 0.10994930 0.097087298 0.085484184
#> CMC36 0.15590661 0.15560744 0.15460520 0.15292873 0.150616797 0.147716746
#> CMC37 0.12521850 0.10710621 0.09138104 0.07779582 0.066108289 0.056088770
#> CMC38 0.06910859 0.06661918 0.06408518 0.06152443 0.058953486 0.056387473
#> CFR1  0.17665164 0.14846984 0.12442185 0.10401421 0.086775543 0.072269689
#> CFR2  0.43933809 0.42243951 0.40410893 0.38455164 0.364012800 0.342764858
#> AMC1  0.30597066 0.28840782 0.26931571 0.24933055 0.229020948 0.208870677
#> AMC2  0.01765646 0.01484091 0.01246956 0.01047376 0.008795013 0.007383667
#> AMC3  0.32180241 0.30009973 0.27762428 0.25495140 0.232568536 0.210868958
#> AMC4  0.11172190 0.10471209 0.09788139 0.09126993 0.084909401 0.078823718
#> AMC5  0.10617989 0.11470474 0.12317257 0.13146278 0.139449394 0.147004655
#> AMC6  0.67694680 0.76739149 0.84142048 0.89239403 0.916037374 0.911176815
#> AMC7  0.16853435 0.15506776 0.14211328 0.12977134 0.118112494 0.107181202
#> AMC8  0.24897222 0.22111182 0.19516210 0.17132881 0.149697436 0.130260978
#> AMC9  0.22330081 0.20856069 0.19383886 0.17933906 0.165230193 0.151646834
#> AMC10 0.36994609 0.37318340 0.37277630 0.36881165 0.361498064 0.351149043
#> AMC11 0.04027012 0.03406947 0.02879404 0.02431443 0.020516735 0.017301551
#> AMC12 0.07645122 0.07114361 0.06606313 0.06122436 0.056636609 0.052304623
#> AFR1  0.38070218 0.37240878 0.36300830 0.35252654 0.341015646 0.328554478
#> AFR2  0.19373274 0.17703585 0.16125936 0.14645441 0.132647523 0.119843972
#> AFR3  0.21972462 0.21540007 0.21068327 0.20559014 0.200141842 0.194364635
#>          theta.40    theta.41
#> CMC1  0.088001739 0.087792065
#> CMC2  0.010441245 0.008627420
#> CMC3  0.120083686 0.113651005
#> CMC4  0.064758013 0.059346508
#> CMC5  0.077104827 0.072632492
#> CMC6  0.245264203 0.214128796
#> CMC7  0.146719459 0.142127803
#> CMC8  0.122605417 0.118445871
#> CMC9  0.101321613 0.098369194
#> CMC10 0.111426462 0.097333233
#> CMC11 0.067478181 0.062129374
#> CMC12 0.152456556 0.148719703
#> CMC13 0.351519967 0.336439761
#> CMC14 0.109146623 0.095906815
#> CMC15 0.082946648 0.073989048
#> CMC16 0.065979255 0.054112767
#> CMC17 0.082593896 0.072959430
#> CMC18 0.357919450 0.331213867
#> CMC19 0.005225385 0.004149084
#> CMC20 0.009729054 0.008431109
#> CMC21 0.014891681 0.012680243
#> CMC22 0.050973600 0.047785396
#> CMC23 0.079034565 0.073254728
#> CMC24 0.199880551 0.199975546
#> CMC25 0.028877259 0.026990198
#> CMC26 0.015994317 0.014392091
#> CMC27 0.110662141 0.100915214
#> CMC28 0.049212653 0.039905764
#> CMC29 0.018977375 0.016760897
#> CMC30 0.176248303 0.163070325
#> CMC31 0.118938938 0.114721559
#> CMC32 0.023798381 0.020527393
#> CMC33 0.021339779 0.018873966
#> CMC34 0.123165225 0.110845322
#> CMC35 0.075078451 0.065794108
#> CMC36 0.144282822 0.140374473
#> CMC37 0.047524737 0.040223106
#> CMC38 0.053840070 0.051323480
#> CFR1  0.060102477 0.049924039
#> CFR2  0.321093629 0.299284433
#> AMC1  0.189271373 0.170523091
#> AMC2  0.006197623 0.005201264
#> AMC3  0.190152949 0.170634473
#> AMC4  0.073029788 0.067538324
#> AMC5  0.154002961 0.160325021
#> AMC6  0.879793732 0.826418106
#> AMC7  0.096999742 0.087572116
#> AMC8  0.112945212 0.097630319
#> AMC9  0.138691128 0.126435706
#> AMC10 0.338159797 0.322980607
#> AMC11 0.014582650 0.012285645
#> AMC12 0.048229281 0.044408242
#> AFR1  0.315247081 0.301219397
#> AFR2  0.108031342 0.097183033
#> AFR3  0.188289550 0.181951831
#> 
#> $tif
#>  [1]  5.145467  5.693289  6.272480  6.879953  7.510673  8.157209  8.809687
#>  [8]  9.456328 10.084589 10.682689 11.241087 11.753403 12.216403 12.629012
#> [15] 12.990712 13.299922 13.553002 13.744267 13.867007 13.915174 13.885162
#> [22] 13.777111 13.595378 13.348156 13.046419 12.702590 12.329203 11.937798
#> [29] 11.538064 11.137245 10.739737 10.346925  9.957284  9.566877  9.170268
#> [36]  8.761805  8.337037  7.893944  7.433623  6.960263  6.480435
#> 
#> $theta
#>  [1] -2.0 -1.9 -1.8 -1.7 -1.6 -1.5 -1.4 -1.3 -1.2 -1.1 -1.0 -0.9 -0.8 -0.7 -0.6
#> [16] -0.5 -0.4 -0.3 -0.2 -0.1  0.0  0.1  0.2  0.3  0.4  0.5  0.6  0.7  0.8  0.9
#> [31]  1.0  1.1  1.2  1.3  1.4  1.5  1.6  1.7  1.8  1.9  2.0
#> 
#> attr(,"class")
#> [1] "info"