Lord-Wingersky Recursion Formula

This function computes the conditional distributions of number-correct (or observed) scores given either the probabilities of category responses for each item or a set of theta values, using the Lord and Wingersky recursion formula (1984).

Usage

lwrc(x = NULL, theta, prob = NULL, cats, D = 1)

Arguments

x

A data frame containing item metadata (e.g., item parameters, number of categories, IRT model types, etc.). See est_irt() or simdat() for more details about the item metadata. This data frame can be easily created using the shape_df() function. If prob is NULL, the item metadata in x is used in the recursion formula. See Details below.

theta

A vector of theta values at which the conditional distributions of observed scores are computed. This argument is required only when item metadata is provided via the x argument.

prob

A matrix containing the category response probabilities for each item. Each row corresponds to an item, and each column represents a score category. If items have different numbers of categories, empty cells should be filled with zeros or NA values. If x is NULL, this matrix is used in the recursion formula.

cats

A numeric vector specifying the number of score categories for each item. For example, a dichotomous item has two categories. This argument is required only when a probability matrix is provided via the prob argument.

D

A scaling constant used in IRT models to make the logistic function closely approximate the normal ogive function. A value of 1.7 is commonly used for this purpose. Default is 1.

Value

When the prob argument is provided, the function returns a vector of probabilities for all possible observed scores on a test.

When the x argument is specified, it returns a matrix of conditional probabilities for each observed score across all specified theta values.

Details

The Lord and Wingersky recursive algorithm provides an efficient method for calculating the compound probabilities of all possible number-correct (i.e., observed) scores on a test, based on IRT models. This algorithm is particularly useful for obtaining the IRT model-based distribution of observed scores.

The conditional distributions of observed scores can be computed using either the item metadata specified in x or the category probability matrix specified in prob.

References

Kolen, M. J. & Brennan, R. L. (2004) Test Equating, Scaling, and Linking (2nd ed.). New York: Springer.

Lord, F. & Wingersky, M. (1984). Comparison of IRT true score and equipercentile observed score equatings. Applied Psychological Measurement, 8(4), 453-461.

Author

Hwanggyu Lim hglim83@gmail.com

Examples

## Example 1: Using a matrix of category probabilities
## This example is from Kolen and Brennan (2004, p. 183)
# Create a matrix of probabilities for correct and incorrect responses to three items
probs <- matrix(c(.74, .73, .82, .26, .27, .18), nrow = 3, ncol = 2, byrow = FALSE)

# Create a vector specifying the number of score categories for each item
cats <- c(2, 2, 2)

# Compute the conditional distribution of observed scores
lwrc(prob = probs, cats = cats)
#>  score.0  score.1  score.2  score.3 
#> 0.442964 0.416708 0.127692 0.012636 

## Example 2: Using a matrix of category probabilities for a mixed-format test
# Category probabilities for a dichotomous item
p1 <- c(0.2, 0.8, 0, 0, 0)

# Category probabilities for another dichotomous item
p2 <- c(0.4, 0.6, NA, NA, NA)

# Category probabilities for a polytomous item with five categories
p3 <- c(0.1, 0.2, 0.2, 0.4, 0.1)

# Category probabilities for a polytomous item with three categories
p4 <- c(0.5, 0.3, 0.2, NA, NA)

# Combine the probability vectors into a matrix
p <- rbind(p1, p2, p3, p4)

# Create a vector specifying the number of score categories for each item
cats <- c(2, 2, 5, 3)

# Compute the conditional distribution of observed scores
lwrc(prob = p, cats = cats)
#> score.0 score.1 score.2 score.3 score.4 score.5 score.6 score.7 score.8 
#>  0.0040  0.0324  0.0956  0.1656  0.2352  0.2452  0.1508  0.0616  0.0096 

## Example 3: Using a data frame of item metadata for a mixed-format test
# Import the "-prm.txt" output file from flexMIRT
flex_prm <- system.file("extdata", "flexmirt_sample-prm.txt", package = "irtQ")

# Read item parameters and convert them to item metadata
x <- bring.flexmirt(file = flex_prm, "par")$Group1$full_df

# Compute the conditional distribution of observed scores for a range of theta values
lwrc(x = x, theta = seq(-4, 4, 0.2), D = 1)
#>               theta.1      theta.2      theta.3      theta.4      theta.5
#> score.0  2.704369e-05 2.137940e-05 1.596819e-05 1.110776e-05 7.068521e-06
#> score.1  3.188232e-04 2.577769e-04 1.978774e-04 1.423061e-04 9.428617e-05
#> score.2  1.836349e-03 1.518756e-03 1.198429e-03 8.912054e-04 6.148972e-04
#> score.3  6.888491e-03 5.828909e-03 4.729158e-03 3.637403e-03 2.614076e-03
#> score.4  1.892929e-02 1.639237e-02 1.367839e-02 1.088468e-02 8.150311e-03
#> score.5  4.064164e-02 3.603004e-02 3.093208e-02 2.547567e-02 1.988283e-02
#> score.6  7.101704e-02 6.447898e-02 5.697778e-02 4.859131e-02 3.954620e-02
#> score.7  1.038959e-01 9.665740e-02 8.796337e-02 7.772070e-02 6.599607e-02
#> score.8  1.299459e-01 1.239515e-01 1.162484e-01 1.064879e-01 9.440701e-02
#> score.9  1.412267e-01 1.382266e-01 1.337044e-01 1.270837e-01 1.177205e-01
#> score.10 1.351424e-01 1.358492e-01 1.356586e-01 1.339166e-01 1.297292e-01
#> score.11 1.151387e-01 1.190036e-01 1.228216e-01 1.260592e-01 1.278329e-01
#> score.12 8.819202e-02 9.384349e-02 1.002303e-01 1.070855e-01 1.137930e-01
#> score.13 6.126554e-02 6.721475e-02 7.439602e-02 8.284495e-02 9.234837e-02
#> score.14 3.891160e-02 4.408546e-02 5.064247e-02 5.885408e-02 6.889133e-02
#> score.15 2.276514e-02 2.667948e-02 3.185495e-02 3.868318e-02 4.759266e-02
#> score.16 1.235405e-02 1.500098e-02 1.864273e-02 2.368224e-02 3.064865e-02
#> score.17 6.258281e-03 7.885490e-03 1.021291e-02 1.358434e-02 1.850385e-02
#> score.18 2.976193e-03 3.896467e-03 5.264680e-03 7.337638e-03 1.052457e-02
#> score.19 1.335163e-03 1.818230e-03 2.564969e-03 3.748014e-03 5.662391e-03
#> score.20 5.673134e-04 8.042811e-04 1.185341e-03 1.816660e-03 2.891392e-03
#> score.21 2.290549e-04 3.382831e-04 5.211079e-04 8.379276e-04 1.405189e-03
#> score.22 8.810854e-05 1.356270e-04 2.184643e-04 3.686578e-04 6.514718e-04
#> score.23 3.235785e-05 5.193941e-05 8.751326e-05 1.550190e-04 2.886988e-04
#> score.24 1.136557e-05 1.903186e-05 3.355393e-05 6.240483e-05 1.224913e-04
#> score.25 3.823835e-06 6.682363e-06 1.233127e-05 2.408426e-05 4.982906e-05
#> score.26 1.233801e-06 2.250965e-06 4.348914e-06 8.921486e-06 1.945724e-05
#> score.27 3.821849e-07 7.281625e-07 1.473283e-06 3.175049e-06 7.299971e-06
#> score.28 1.137478e-07 2.263924e-07 4.798176e-07 1.086493e-06 2.633641e-06
#> score.29 3.254982e-08 6.769680e-08 1.503330e-07 3.577430e-07 9.143016e-07
#> score.30 8.960809e-09 1.948104e-08 4.534069e-08 1.134089e-07 3.056148e-07
#> score.31 2.374509e-09 5.398000e-09 1.317080e-08 3.463195e-08 9.840504e-08
#> score.32 6.059608e-10 1.440920e-09 3.686552e-09 1.019140e-08 3.053298e-08
#> score.33 1.489858e-10 3.706804e-10 9.946145e-10 2.890907e-09 9.131088e-09
#> score.34 3.530306e-11 9.192276e-11 2.587017e-10 7.905621e-10 2.632148e-09
#> score.35 8.063550e-12 2.197668e-11 6.487537e-11 2.084199e-10 7.313197e-10
#> score.36 1.775443e-12 5.065395e-12 1.568440e-11 5.296492e-11 1.958075e-10
#> score.37 3.768102e-13 1.125442e-12 3.654957e-12 1.297086e-11 5.050425e-11
#> score.38 7.707194e-14 2.409829e-13 8.206952e-13 3.059849e-12 1.254246e-11
#> score.39 1.518800e-14 4.970927e-14 1.774825e-13 6.949015e-13 2.997033e-12
#> score.40 2.882336e-15 9.872690e-15 3.694104e-14 1.518077e-13 6.884378e-13
#> score.41 5.264488e-16 1.886478e-15 7.393645e-15 3.186952e-14 1.518515e-13
#> score.42 9.246364e-17 3.464697e-16 1.421434e-15 6.421503e-15 3.211973e-14
#> score.43 1.560031e-17 6.108902e-17 2.621443e-16 1.240056e-15 6.504700e-15
#> score.44 2.525221e-18 1.032615e-17 4.630404e-17 2.291038e-16 1.258807e-15
#> score.45 3.915940e-19 1.670605e-18 7.818985e-18 4.041169e-17 2.322680e-16
#> score.46 5.807630e-20 2.581732e-19 1.259394e-18 6.788809e-18 4.075440e-17
#> score.47 8.220369e-21 3.801990e-20 1.929690e-19 1.083017e-18 6.779571e-18
#> score.48 1.107675e-21 5.320069e-21 2.803872e-20 1.635283e-19 1.065629e-18
#> score.49 1.416497e-22 7.049259e-22 3.849506e-21 2.328456e-20 1.576857e-19
#> score.50 1.712745e-23 8.810176e-23 4.973822e-22 3.114117e-21 2.188132e-20
#> score.51 1.949805e-24 1.034079e-23 6.022099e-23 3.895646e-22 2.835960e-21
#> score.52 2.079982e-25 1.134569e-24 6.801821e-24 4.538692e-23 3.418858e-22
#> score.53 2.068725e-26 1.157991e-25 7.133670e-25 4.903179e-24 3.817612e-23
#> score.54 1.908220e-27 1.093973e-26 6.914459e-26 4.889566e-25 3.931574e-24
#> score.55 1.623583e-28 9.517234e-28 6.163910e-27 4.480247e-26 3.717704e-25
#> score.56 1.267094e-29 7.584417e-29 5.028328e-28 3.753846e-27 3.212817e-26
#> score.57 9.017778e-31 5.505983e-30 3.733806e-29 2.861271e-28 2.524754e-27
#> score.58 5.816496e-32 3.619685e-31 2.509248e-30 1.972931e-29 1.794231e-28
#> score.59 3.377327e-33 2.140899e-32 1.516458e-31 1.222960e-30 1.145968e-29
#> score.60 1.752135e-34 1.130870e-33 8.182064e-33 6.766220e-32 6.531458e-31
#> score.61 8.051648e-36 5.289479e-35 3.908138e-34 3.313341e-33 3.294274e-32
#> score.62 3.243933e-37 2.168607e-36 1.635918e-35 1.421672e-34 1.455652e-33
#> score.63 1.131611e-38 7.696815e-38 5.927198e-37 5.279196e-36 5.565839e-35
#> score.64 3.364515e-40 2.327999e-39 1.829883e-38 1.670191e-37 1.812904e-36
#> score.65 8.352608e-42 5.878684e-41 4.715985e-40 4.410474e-39 4.928090e-38
#> score.66 1.683480e-43 1.205091e-42 9.865405e-42 9.452389e-41 1.087064e-39
#> score.67 2.644897e-45 1.925445e-44 1.608341e-43 1.578551e-42 1.868192e-41
#> score.68 3.037089e-47 2.248253e-46 1.915972e-45 1.926005e-44 2.345262e-43
#> score.69 2.266019e-49 1.705557e-48 1.482677e-47 1.526262e-46 1.911823e-45
#> score.70 8.240346e-52 6.305324e-51 5.590563e-50 5.892107e-49 7.590610e-48
#>               theta.6      theta.7      theta.8      theta.9     theta.10
#> score.0  4.024277e-06 1.994297e-06 8.320742e-07 2.809094e-07 7.327976e-08
#> score.1  5.635870e-05 2.961165e-05 1.324716e-05 4.856650e-06 1.395288e-06
#> score.2  3.859519e-04 2.150166e-04 1.031370e-04 4.105533e-05 1.298529e-05
#> score.3  1.723280e-03 1.018111e-03 5.236475e-04 2.263023e-04 7.877559e-05
#> score.4  5.644647e-03 3.537260e-03 1.950996e-03 9.153267e-04 3.505754e-04
#> score.5  1.447155e-02 9.621777e-03 5.691917e-03 2.898956e-03 1.221360e-03
#> score.6  3.026212e-02 2.135496e-02 1.355211e-02 7.493097e-03 3.471851e-03
#> score.7  5.312403e-02 3.980405e-02 2.710505e-02 1.627021e-02 8.288771e-03
#> score.8  7.998597e-02 6.366361e-02 4.653216e-02 3.032510e-02 1.698187e-02
#> score.9  1.050492e-01 8.886639e-02 6.973781e-02 4.934370e-02 3.036467e-02
#> score.10 1.220211e-01 1.097706e-01 9.251458e-02 7.106814e-02 4.803951e-02
#> score.11 1.268361e-01 1.214060e-01 1.099167e-01 9.166051e-02 6.802681e-02
#> score.12 1.191989e-01 1.214619e-01 1.181507e-01 1.069334e-01 8.707959e-02
#> score.13 1.022074e-01 1.109209e-01 1.159330e-01 1.138389e-01 1.016408e-01
#> score.14 8.061566e-02 9.320880e-02 1.046685e-01 1.114530e-01 1.090058e-01
#> score.15 5.891797e-02 7.258967e-02 8.755987e-02 1.010412e-01 1.081402e-01
#> score.16 4.015657e-02 5.272204e-02 6.828757e-02 8.533521e-02 9.982843e-02
#> score.17 2.566613e-02 3.590587e-02 4.991554e-02 6.749248e-02 8.619809e-02
#> score.18 1.545647e-02 2.303570e-02 3.435297e-02 5.021550e-02 6.993027e-02
#> score.19 8.804914e-03 1.397627e-02 2.234624e-02 3.528172e-02 5.351008e-02
#> score.20 4.760329e-03 8.045549e-03 1.378404e-02 2.348657e-02 3.874824e-02
#> score.21 2.449311e-03 4.406478e-03 8.085088e-03 1.485480e-02 2.662902e-02
#> score.22 1.202142e-03 2.301512e-03 4.520171e-03 8.948182e-03 1.741056e-02
#> score.23 5.639334e-04 1.148634e-03 2.413569e-03 5.144170e-03 1.085281e-02
#> score.24 2.532695e-04 5.486875e-04 1.232931e-03 2.827266e-03 6.461421e-03
#> score.25 1.090503e-04 2.512202e-04 6.034131e-04 1.487749e-03 3.679911e-03
#> score.26 4.506770e-05 1.103773e-04 2.832740e-04 7.504810e-04 2.007384e-03
#> score.27 1.789449e-05 4.658279e-05 1.276869e-04 3.632803e-04 1.049964e-03
#> score.28 6.831907e-06 1.889924e-05 5.530815e-05 1.688889e-04 5.270499e-04
#> score.29 2.509747e-06 7.376118e-06 2.303695e-05 7.545918e-05 2.540804e-04
#> score.30 8.876186e-07 2.770823e-06 9.231685e-06 3.241928e-05 1.176988e-04
#> score.31 3.023572e-07 1.002218e-06 3.560607e-06 1.339804e-05 5.241178e-05
#> score.32 9.922991e-08 3.491430e-07 1.322087e-06 5.327578e-06 2.244157e-05
#> score.33 3.138083e-08 1.171621e-07 4.726431e-07 2.038503e-06 9.240458e-06
#> score.34 9.563066e-09 3.787102e-08 1.626773e-07 7.505263e-07 3.658772e-06
#> score.35 2.807967e-09 1.178950e-08 5.389649e-08 2.658326e-07 1.392833e-06
#> score.36 7.942143e-10 3.533613e-09 1.718248e-08 9.054938e-08 5.096076e-07
#> score.37 2.162992e-10 1.019233e-09 5.268457e-09 2.964638e-08 1.791121e-07
#> score.38 5.668766e-11 2.827338e-10 1.552582e-09 9.323152e-09 6.043200e-08
#> score.39 1.428565e-11 7.536429e-11 4.393569e-10 2.813655e-09 1.955599e-08
#> score.40 3.458297e-12 1.928333e-11 1.192611e-10 8.139869e-10 6.063051e-09
#> score.41 8.032533e-13 4.730144e-12 3.101188e-11 2.254372e-10 1.798587e-09
#> score.42 1.787490e-13 1.110662e-12 7.713019e-12 5.967734e-11 5.097102e-10
#> score.43 3.804433e-14 2.491857e-13 1.831413e-12 1.507175e-11 1.377440e-10
#> score.44 7.728773e-15 5.330662e-14 4.142669e-13 3.623725e-12 3.542105e-11
#> score.45 1.495115e-15 1.084664e-14 8.905048e-14 8.274099e-13 8.646602e-12
#> score.46 2.746602e-16 2.093446e-15 1.814059e-14 1.789253e-13 1.998296e-12
#> score.47 4.776809e-17 3.820708e-16 3.491434e-15 3.653497e-14 4.359450e-13
#> score.48 7.838549e-18 6.571960e-17 6.328102e-16 7.021671e-15 8.949474e-14
#> score.49 1.209267e-18 1.061656e-17 1.076380e-16 1.265916e-15 1.723145e-14
#> score.50 1.747266e-19 1.604799e-18 1.712113e-17 2.133474e-16 3.101085e-15
#> score.51 2.355372e-20 2.261382e-19 2.537387e-18 3.349130e-17 5.198000e-16
#> score.52 2.950550e-21 2.959236e-20 3.490649e-19 4.879145e-18 8.085528e-17
#> score.53 3.420860e-22 3.582094e-21 4.440447e-20 6.571741e-19 1.162778e-17
#> score.54 3.655572e-23 3.994771e-22 5.202656e-21 8.151467e-20 1.539925e-18
#> score.55 3.585004e-24 4.087050e-23 5.591022e-22 9.272797e-21 1.870325e-19
#> score.56 3.211842e-25 3.818899e-24 5.486433e-23 9.631127e-22 2.074047e-20
#> score.57 2.615801e-26 3.243068e-25 4.892300e-24 9.089270e-23 2.089743e-21
#> score.58 1.926079e-27 2.489503e-26 3.942908e-25 7.752155e-24 1.902780e-22
#> score.59 1.274356e-28 1.716910e-27 2.854591e-26 5.938786e-25 1.556103e-23
#> score.60 7.522757e-30 1.056304e-28 1.843415e-27 4.057684e-26 1.134899e-24
#> score.61 3.929247e-31 5.749313e-30 1.053005e-28 2.452081e-27 7.319928e-26
#> score.62 1.797745e-32 2.740740e-31 5.267470e-30 1.297466e-28 4.133398e-27
#> score.63 7.116398e-34 1.130235e-32 2.279059e-31 5.937071e-30 2.018185e-28
#> score.64 2.399383e-35 3.969242e-34 8.396017e-33 2.312800e-31 8.387527e-30
#> score.65 6.750418e-37 1.162953e-35 2.580034e-34 7.513729e-33 2.906571e-31
#> score.66 1.540848e-38 2.763960e-37 6.429874e-36 1.979277e-34 8.165387e-33
#> score.67 2.739663e-40 5.115845e-39 1.247661e-37 4.058593e-36 1.785251e-34
#> score.68 3.557522e-42 6.913762e-41 1.767233e-39 6.073502e-38 2.847868e-36
#> score.69 2.999051e-44 6.064363e-43 1.624224e-41 5.895782e-40 2.946302e-38
#> score.70 1.231066e-46 2.589365e-45 7.264500e-44 2.784371e-42 1.482561e-40
#>              theta.11     theta.12     theta.13     theta.14     theta.15
#> score.0  1.402403e-08 1.860825e-09 1.613483e-10 8.613263e-12 2.673977e-13
#> score.1  2.985585e-07 4.500083e-08 4.504642e-09 2.821368e-10 1.044160e-11
#> score.2  3.104793e-06 5.311208e-07 6.130298e-08 4.497655e-09 1.980369e-10
#> score.3  2.103499e-05 4.080340e-06 5.423882e-07 4.654119e-08 2.433287e-09
#> score.4  1.044884e-04 2.296397e-05 3.511312e-06 3.518371e-07 2.179929e-08
#> score.5  4.061060e-04 1.010362e-04 1.774970e-05 2.073630e-06 1.519598e-07
#> score.6  1.287176e-03 3.622126e-04 7.302060e-05 9.930515e-06 8.590440e-07
#> score.7  3.424617e-03 1.089033e-03 2.516246e-04 3.977145e-05 4.053261e-06
#> score.8  7.814421e-03 2.805573e-03 7.419988e-04 1.360820e-04 1.630624e-05
#> score.9  1.555191e-02 6.297429e-03 1.903801e-03 4.044437e-04 5.686479e-05
#> score.10 2.736471e-02 1.248361e-02 4.307658e-03 1.058140e-03 1.742004e-04
#> score.11 4.305964e-02 2.210307e-02 8.691900e-03 2.464160e-03 4.739739e-04
#> score.12 6.118801e-02 3.529281e-02 1.578974e-02 5.156138e-03 1.156161e-03
#> score.13 7.919109e-02 5.124824e-02 2.603798e-02 9.773531e-03 2.548882e-03
#> score.14 9.404909e-02 6.817565e-02 3.926093e-02 1.690279e-02 5.114798e-03
#> score.15 1.031752e-01 8.363051e-02 5.447805e-02 2.684101e-02 9.401232e-03
#> score.16 1.051657e-01 9.514534e-02 6.996228e-02 3.935712e-02 1.591674e-02
#> score.17 1.001098e-01 1.009034e-01 8.357602e-02 5.355687e-02 2.494679e-02
#> score.18 8.939751e-02 1.001984e-01 9.328584e-02 6.793864e-02 3.635920e-02
#> score.19 7.518111e-02 9.353023e-02 9.767359e-02 8.065884e-02 4.947561e-02
#> score.20 5.974324e-02 8.234979e-02 9.626551e-02 8.993852e-02 6.308009e-02
#> score.21 4.499165e-02 6.859403e-02 8.958134e-02 9.448124e-02 7.559469e-02
#> score.22 3.219103e-02 5.419408e-02 7.891827e-02 9.376507e-02 8.538987e-02
#> score.23 2.193040e-02 4.070452e-02 6.597343e-02 8.812167e-02 9.114177e-02
#> score.24 1.425230e-02 2.912153e-02 5.244272e-02 7.859562e-02 9.212619e-02
#> score.25 8.850147e-03 1.987954e-02 3.971065e-02 6.665104e-02 8.835971e-02
#> score.26 5.258205e-03 1.296743e-02 2.868882e-02 5.383053e-02 8.055386e-02
#> score.27 2.992568e-03 8.092732e-03 1.980097e-02 4.146608e-02 6.991140e-02
#> score.28 1.632981e-03 4.837038e-03 1.307146e-02 3.050311e-02 5.783990e-02
#> score.29 8.550312e-04 2.771247e-03 8.261102e-03 2.145089e-02 4.567064e-02
#> score.30 4.298399e-04 1.522912e-03 5.002233e-03 1.443386e-02 3.445205e-02
#> score.31 2.075623e-04 8.031571e-04 2.903779e-03 9.299664e-03 2.485006e-02
#> score.32 9.630197e-05 4.066369e-04 1.616699e-03 5.740359e-03 1.715020e-02
#> score.33 4.293696e-05 1.976895e-04 8.635471e-04 3.396006e-03 1.133092e-02
#> score.34 1.839659e-05 9.229030e-05 4.425812e-04 1.926009e-03 7.169214e-03
#> score.35 7.573355e-06 4.136980e-05 2.176422e-04 1.047235e-03 4.344858e-03
#> score.36 2.994681e-06 1.780151e-05 1.026742e-04 5.458765e-04 2.522300e-03
#> score.37 1.136889e-06 7.350134e-06 4.645175e-05 2.727134e-04 1.402441e-03
#> score.38 4.141021e-07 2.910327e-06 2.014392e-05 1.305282e-04 7.466488e-04
#> score.39 1.445950e-07 1.104227e-06 8.367316e-06 5.981827e-05 3.804454e-04
#> score.40 4.835052e-08 4.010693e-07 3.326157e-06 2.622751e-05 1.854094e-04
#> score.41 1.546326e-08 1.392856e-07 1.263968e-06 1.099125e-05 8.635083e-05
#> score.42 4.722781e-09 4.618461e-08 4.585525e-07 4.397248e-06 3.839151e-05
#> score.43 1.375060e-09 1.459687e-08 1.585684e-07 1.676989e-06 1.627328e-05
#> score.44 3.808748e-10 4.388759e-09 5.216949e-08 6.086296e-07 6.566181e-06
#> score.45 1.001302e-10 1.252474e-09 1.629537e-08 2.097902e-07 2.517418e-06
#> score.46 2.491929e-11 3.384110e-10 4.820706e-09 6.852284e-08 9.151265e-07
#> score.47 5.853977e-12 8.632944e-11 1.347056e-09 2.115361e-08 3.146553e-07
#> score.48 1.294117e-12 2.073009e-11 3.544940e-10 6.154473e-09 1.020536e-07
#> score.49 2.683422e-13 4.670624e-12 8.757939e-11 1.682289e-09 3.112748e-08
#> score.50 5.201413e-14 9.840382e-13 2.024429e-11 4.305859e-10 8.899097e-09
#> score.51 9.391648e-15 1.931894e-13 4.362902e-12 1.028315e-10 2.376241e-09
#> score.52 1.573880e-15 3.521293e-14 8.734093e-13 2.282843e-11 5.903848e-10
#> score.53 2.438789e-16 5.936334e-15 1.617918e-13 4.692528e-12 1.359389e-10
#> score.54 3.480488e-17 9.219439e-16 2.762084e-14 8.894618e-13 2.888563e-11
#> score.55 4.555758e-18 1.313506e-16 4.327138e-15 1.547890e-13 5.639023e-12
#> score.56 5.444930e-19 1.708982e-17 6.192345e-16 2.461574e-14 1.006530e-12
#> score.57 5.913048e-20 2.020583e-18 8.054303e-17 3.559065e-15 1.634151e-13
#> score.58 5.803017e-21 2.159072e-19 9.469152e-18 4.652293e-16 2.399452e-14
#> score.59 5.114929e-22 2.072104e-20 9.999644e-19 5.463314e-17 3.165893e-15
#> score.60 4.020424e-23 1.773363e-21 9.417098e-20 5.721981e-18 3.726080e-16
#> score.61 2.794484e-24 1.342036e-22 7.842073e-21 5.299547e-19 3.878390e-17
#> score.62 1.700353e-25 8.890167e-24 5.716267e-22 4.296385e-20 3.533782e-18
#> score.63 8.944953e-27 5.091200e-25 3.601949e-23 3.010944e-21 2.783305e-19
#> score.64 4.004767e-28 2.481117e-26 1.931309e-24 1.795453e-22 1.865251e-20
#> score.65 1.494807e-29 1.007941e-27 8.631679e-26 8.923853e-24 1.041822e-21
#> score.66 4.522432e-31 3.318574e-29 3.126335e-27 3.594205e-25 4.715071e-23
#> score.67 1.064663e-32 8.501018e-31 8.809437e-29 1.126165e-26 1.659946e-24
#> score.68 1.828415e-34 1.588419e-32 1.810547e-30 2.573535e-28 4.261768e-26
#> score.69 2.036101e-36 1.924322e-34 2.412522e-32 3.812797e-30 7.093058e-28
#> score.70 1.102628e-38 1.133600e-36 1.563126e-34 2.746707e-32 5.739806e-30
#>              theta.16     theta.17     theta.18     theta.19     theta.20
#> score.0  4.587477e-15 4.170205e-17 1.946316e-19 4.566616e-22 5.327930e-25
#> score.1  2.169265e-13 2.425207e-15 1.413515e-17 4.204417e-20 6.310127e-23
#> score.2  4.970095e-12 6.813818e-14 4.942643e-16 1.856492e-18 3.568440e-21
#> score.3  7.359300e-11 1.233658e-12 1.109953e-14 5.244077e-17 1.285253e-19
#> score.4  7.926265e-10 1.619975e-11 1.801712e-13 1.066552e-15 3.318293e-18
#> score.5  6.626753e-09 1.646554e-10 2.256060e-12 1.666787e-14 6.553990e-17
#> score.6  4.482280e-08 1.350109e-09 2.271298e-11 2.086125e-13 1.032147e-15
#> score.7  2.524433e-07 9.191574e-09 1.892190e-10 2.152162e-12 1.333957e-14
#> score.8  1.209339e-06 5.307551e-08 1.332546e-09 1.869611e-11 1.445383e-13
#> score.9  5.009833e-06 2.642734e-07 8.065015e-09 1.390456e-10 1.334954e-12
#> score.10 1.818665e-05 1.149820e-06 4.251120e-08 8.971712e-10 1.065100e-11
#> score.11 5.849309e-05 4.419670e-06 1.973117e-07 5.078067e-09 7.422787e-11
#> score.12 1.682366e-04 1.514862e-05 8.139641e-07 2.545065e-08 4.561338e-10
#> score.13 4.362074e-04 4.667365e-05 3.008625e-06 1.138702e-07 2.491890e-09
#> score.14 1.026802e-03 1.301826e-04 1.003525e-05 4.580916e-07 1.219080e-08
#> score.15 2.208115e-03 3.307829e-04 3.039750e-05 1.667700e-06 5.375767e-08
#> score.16 4.362444e-03 7.699929e-04 8.409553e-05 5.526265e-06 2.149483e-07
#> score.17 7.957819e-03 1.650387e-03 2.135860e-04 1.675678e-05 7.835701e-07
#> score.18 1.346402e-02 3.272096e-03 5.003432e-04 4.671861e-05 2.617263e-06
#> score.19 2.121441e-02 6.025569e-03 1.085668e-03 1.202926e-04 8.047148e-06
#> score.20 3.124176e-02 1.034453e-02 2.190390e-03 2.871918e-04 2.287144e-05
#> score.21 4.314118e-02 1.661139e-02 4.123199e-03 6.380526e-04 6.032101e-05
#> score.22 5.602082e-02 2.502479e-02 7.263965e-03 1.323427e-03 1.481397e-04
#> score.23 6.858325e-02 3.546093e-02 1.200972e-02 2.570159e-03 3.398201e-04
#> score.24 7.933862e-02 4.737663e-02 1.867980e-02 4.685485e-03 7.301269e-04
#> score.25 8.690160e-02 5.980244e-02 2.739265e-02 8.036628e-03 1.472913e-03
#> score.26 9.028807e-02 7.145319e-02 3.794493e-02 1.299552e-02 2.795873e-03
#> score.27 8.912320e-02 8.094552e-02 4.973640e-02 1.984669e-02 5.003085e-03
#> score.28 8.370106e-02 8.707167e-02 6.178136e-02 2.867084e-02 8.453909e-03
#> score.29 7.488629e-02 8.905342e-02 7.282736e-02 3.923351e-02 1.350851e-02
#> score.30 6.389824e-02 8.670200e-02 8.156735e-02 5.091869e-02 2.043814e-02
#> score.31 5.204852e-02 8.043940e-02 8.689591e-02 6.274556e-02 2.931213e-02
#> score.32 4.050531e-02 7.118161e-02 8.813932e-02 7.348668e-02 3.988949e-02
#> score.33 3.013587e-02 6.012632e-02 8.519300e-02 8.187380e-02 5.155399e-02
#> score.34 2.144566e-02 4.851046e-02 7.852836e-02 8.684439e-02 6.332991e-02
#> score.35 1.460249e-02 3.740171e-02 6.907271e-02 8.776187e-02 7.399711e-02
#> score.36 9.515348e-03 2.756621e-02 5.800327e-02 8.454667e-02 8.229296e-02
#> score.37 5.933927e-03 1.942529e-02 4.651666e-02 7.768175e-02 8.715600e-02
#> score.38 3.540926e-03 1.308783e-02 3.563286e-02 6.809551e-02 8.794631e-02
#> score.39 2.021210e-03 8.429626e-03 2.607256e-02 5.696054e-02 8.458022e-02
#> score.40 1.103085e-03 5.188520e-03 1.821952e-02 4.546675e-02 7.754146e-02
#> score.41 5.751856e-04 3.050316e-03 1.215516e-02 3.462660e-02 6.776832e-02
#> score.42 2.862967e-04 1.711573e-03 7.737835e-03 2.515199e-02 5.645264e-02
#> score.43 1.358762e-04 9.157647e-04 4.696626e-03 1.741578e-02 4.480850e-02
#> score.44 6.140339e-05 4.666541e-04 2.715396e-03 1.148650e-02 3.387019e-02
#> score.45 2.637845e-05 2.261527e-04 1.493568e-03 7.208863e-03 2.436226e-02
#> score.46 1.075162e-05 1.040526e-04 7.803809e-04 4.299578e-03 1.665754e-02
#> score.47 4.148481e-06 4.535895e-05 3.866263e-04 2.433231e-03 1.081249e-02
#> score.48 1.511390e-06 1.868958e-05 1.812372e-04 1.304116e-03 6.652031e-03
#> score.49 5.184105e-07 7.258993e-06 8.018331e-05 6.604567e-04 3.871107e-03
#> score.50 1.668707e-07 2.649421e-06 3.338427e-05 3.152198e-04 2.125865e-03
#> score.51 5.023010e-08 9.055792e-07 1.303725e-05 1.413449e-04 1.098575e-03
#> score.52 1.408553e-08 2.887737e-07 4.757802e-06 5.933382e-05 5.324476e-04
#> score.53 3.664718e-09 8.555775e-08 1.615924e-06 2.322383e-05 2.411060e-04
#> score.54 8.808180e-10 2.344820e-08 5.084875e-07 8.437731e-06 1.015578e-04
#> score.55 1.946740e-10 5.916256e-09 1.475263e-07 2.831564e-06 3.959487e-05
#> score.56 3.936967e-11 1.367278e-09 3.925542e-08 8.729299e-07 1.420963e-05
#> score.57 7.246478e-12 2.878347e-10 9.525486e-09 2.457593e-07 4.665388e-06
#> score.58 1.206858e-12 5.486326e-11 2.094668e-09 6.277626e-08 1.391939e-06
#> score.59 1.806772e-13 9.404892e-12 4.145457e-10 1.444479e-08 3.745655e-07
#> score.60 2.413391e-14 1.438969e-12 7.325907e-11 2.969942e-09 9.015008e-08
#> score.61 2.851398e-15 1.947796e-13 1.145696e-11 5.406123e-10 1.922122e-08
#> score.62 2.949196e-16 2.308271e-14 1.568845e-12 8.618072e-11 3.590299e-09
#> score.63 2.636812e-17 2.364561e-15 1.856940e-13 1.187509e-11 5.797151e-10
#> score.64 2.005775e-18 2.060606e-16 1.869508e-14 1.391532e-12 7.958944e-11
#> score.65 1.271506e-19 1.496213e-17 1.567815e-15 1.357806e-13 9.095358e-12
#> score.66 6.530271e-21 8.799612e-19 1.064577e-16 1.072233e-14 8.407106e-13
#> score.67 2.608446e-22 4.023884e-20 5.617989e-18 6.576667e-16 6.031478e-14
#> score.68 7.596985e-24 1.341198e-21 2.159894e-19 2.936780e-17 3.147570e-15
#> score.69 1.434040e-25 2.896306e-23 5.377094e-21 8.485293e-19 1.061770e-16
#> score.70 1.315861e-27 3.039187e-25 6.500783e-23 1.189604e-20 1.736016e-18
#>              theta.21     theta.22     theta.23     theta.24     theta.25
#> score.0  3.090216e-28 9.011119e-32 1.351248e-35 1.075654e-39 4.720494e-44
#> score.1  4.774397e-26 1.840419e-29 3.693881e-33 3.981704e-37 2.391782e-41
#> score.2  3.504786e-24 1.776066e-27 4.741521e-31 6.871661e-35 5.604162e-39
#> score.3  1.630527e-22 1.080273e-25 3.812568e-29 7.378058e-33 8.107049e-37
#> score.4  5.410901e-21 4.661408e-24 2.161849e-27 5.549847e-31 8.157711e-35
#> score.5  1.366926e-19 1.523005e-22 9.227926e-26 3.122991e-29 6.099907e-33
#> score.6  2.739980e-18 3.927500e-21 3.091420e-24 1.371004e-27 3.536208e-31
#> score.7  4.485530e-17 8.228818e-20 8.368259e-23 4.835578e-26 1.637331e-29
#> score.8  6.126893e-16 1.431203e-18 1.870426e-21 1.400582e-24 6.191127e-28
#> score.9  7.099854e-15 2.101202e-17 3.510735e-20 3.388722e-23 1.945248e-26
#> score.10 7.073946e-14 2.639327e-16 5.609431e-19 6.944177e-22 5.150449e-25
#> score.11 6.127902e-13 2.868421e-15 7.716393e-18 1.219100e-20 1.162624e-23
#> score.12 4.659281e-12 2.723089e-14 9.227078e-17 1.851484e-19 2.259692e-22
#> score.13 3.135238e-11 2.276978e-13 9.671640e-16 2.453160e-18 3.814011e-21
#> score.14 1.880838e-10 1.689415e-12 8.952336e-15 2.856835e-17 5.632334e-20
#> score.15 1.012593e-09 1.119659e-11 7.366588e-14 2.943663e-16 7.325893e-19
#> score.16 4.922073e-09 6.668930e-11 5.421708e-13 2.699990e-15 8.443205e-18
#> score.17 2.172280e-08 3.590091e-10 3.589224e-12 2.216833e-14 8.669936e-17
#> score.18 8.749643e-08 1.756041e-09 2.148655e-11 1.637846e-13 7.972680e-16
#> score.19 3.231960e-07 7.843500e-09 1.169023e-10 1.094314e-12 6.597312e-15
#> score.20 1.099727e-06 3.214177e-08 5.808440e-10 6.643777e-12 4.935296e-14
#> score.21 3.461266e-06 1.213737e-07 2.647710e-09 3.682166e-11 3.352732e-13
#> score.22 1.011456e-05 4.240794e-07 1.112117e-08 1.871368e-10 2.077538e-12
#> score.23 2.753534e-05 1.376130e-06 4.321923e-08 8.759260e-10 1.179415e-11
#> score.24 7.004442e-05 4.161225e-06 1.559870e-07 3.791625e-09 6.160678e-11
#> score.25 1.669361e-04 1.176049e-05 5.246411e-07 1.523735e-08 2.973435e-10
#> score.26 3.736212e-04 3.114588e-05 1.649301e-06 5.704817e-08 1.331342e-09
#> score.27 7.868640e-04 7.746803e-05 4.858739e-06 1.995994e-07 5.550260e-09
#> score.28 1.562156e-03 1.813142e-04 1.344270e-05 6.543364e-07 2.161388e-08
#> score.29 2.928033e-03 3.999941e-04 3.499409e-05 2.014248e-06 7.883693e-08
#> score.30 5.188458e-03 8.329422e-04 8.584817e-05 5.832717e-06 2.699353e-07
#> score.31 8.701991e-03 1.639299e-03 1.987382e-04 1.591169e-05 8.691118e-07
#> score.32 1.382795e-02 3.052467e-03 4.346611e-04 4.094404e-05 2.634952e-06
#> score.33 2.083726e-02 5.382624e-03 8.990308e-04 9.948604e-05 7.530777e-06
#> score.34 2.979936e-02 8.995587e-03 1.760042e-03 2.284751e-04 2.030958e-05
#> score.35 4.047262e-02 1.425778e-02 3.263671e-03 4.963318e-04 5.172948e-05
#> score.36 5.223734e-02 2.144472e-02 5.735639e-03 1.020596e-03 1.245354e-04
#> score.37 6.410881e-02 3.062441e-02 9.557990e-03 1.987554e-03 2.835678e-04
#> score.38 7.485055e-02 4.154410e-02 1.510952e-02 3.667352e-03 6.110279e-04
#> score.39 8.317659e-02 5.356013e-02 2.266777e-02 6.413677e-03 1.246455e-03
#> score.40 8.799917e-02 6.565027e-02 3.228537e-02 1.063451e-02 2.407841e-03
#> score.41 8.865714e-02 7.652977e-02 4.367095e-02 1.672301e-02 4.405707e-03
#> score.42 8.506023e-02 8.486164e-02 5.611728e-02 2.494741e-02 7.637239e-03
#> score.43 7.770728e-02 8.951712e-02 6.851816e-02 3.531555e-02 1.254565e-02
#> score.44 6.757395e-02 8.981889e-02 7.949733e-02 4.744855e-02 1.953391e-02
#> score.45 5.590431e-02 8.569686e-02 8.763952e-02 6.051113e-02 2.883442e-02
#> score.46 4.396633e-02 7.770875e-02 9.177559e-02 7.324468e-02 4.035574e-02
#> score.47 3.283576e-02 6.691833e-02 9.124624e-02 8.412708e-02 5.354951e-02
#> score.48 2.325615e-02 5.466735e-02 8.606573e-02 9.164448e-02 6.735418e-02
#> score.49 1.559398e-02 4.230730e-02 7.693215e-02 9.461537e-02 8.026684e-02
#> score.50 9.878781e-03 3.096308e-02 6.507745e-02 9.247758e-02 9.056369e-02
#> score.51 5.897740e-03 2.138305e-02 5.200049e-02 8.544785e-02 9.663957e-02
#> score.52 3.308233e-03 1.389750e-02 3.916047e-02 7.449665e-02 9.738649e-02
#> score.53 1.737331e-03 8.473166e-03 2.771579e-02 6.113700e-02 9.249902e-02
#> score.54 8.505979e-04 4.827412e-03 1.837143e-02 4.708729e-02 8.259994e-02
#> score.55 3.863791e-04 2.558249e-03 1.135700e-02 3.390939e-02 6.912688e-02
#> score.56 1.619367e-04 1.254249e-03 6.514452e-03 2.272752e-02 5.400082e-02
#> score.57 6.223106e-05 5.653583e-04 3.446267e-03 1.409739e-02 3.917847e-02
#> score.58 2.177542e-05 2.326297e-04 1.669484e-03 8.036955e-03 2.623210e-02
#> score.59 6.883950e-06 8.667690e-05 7.345106e-04 4.176792e-03 1.608192e-02
#> score.60 1.949051e-06 2.897886e-05 2.907342e-04 1.959765e-03 8.941269e-03
#> score.61 4.893372e-07 8.604010e-06 1.024204e-04 8.209403e-04 4.457051e-03
#> score.62 1.076941e-07 2.241669e-06 3.171465e-05 3.030566e-04 1.965289e-03
#> score.63 2.049393e-08 5.053099e-07 8.506624e-06 9.710039e-05 7.544507e-04
#> score.64 3.315842e-09 9.686627e-08 1.941676e-06 2.651188e-05 2.473932e-04
#> score.65 4.464192e-10 1.544915e-08 3.688166e-07 6.028808e-06 6.767765e-05
#> score.66 4.858519e-11 1.990887e-09 5.659485e-08 1.107890e-06 1.497744e-05
#> score.67 4.100821e-12 1.988245e-10 6.726591e-09 1.576774e-07 2.568453e-06
#> score.68 2.515269e-13 1.441476e-11 5.799151e-10 1.627024e-08 3.193795e-07
#> score.69 9.960936e-15 6.739339e-13 3.220524e-11 1.080635e-09 2.555687e-08
#> score.70 1.909490e-16 1.523040e-14 8.633509e-13 3.461134e-11 9.857196e-10
#>              theta.26     theta.27     theta.28     theta.29     theta.30
#> score.0  1.189266e-48 1.792503e-53 1.684717e-58 1.030126e-63 4.284125e-69
#> score.1  8.331373e-46 1.752562e-50 2.319135e-55 2.013031e-60 1.197699e-65
#> score.2  2.673794e-43 7.765038e-48 1.428660e-52 1.735102e-57 1.452534e-62
#> score.3  5.252903e-41 2.086272e-45 5.281736e-50 8.874193e-55 1.032606e-59
#> score.4  7.123008e-39 3.836934e-43 1.324888e-47 3.051112e-52 4.887383e-57
#> score.5  7.127042e-37 5.168532e-41 2.415564e-45 7.564856e-50 1.654793e-54
#> score.6  5.492815e-35 5.327096e-39 3.346890e-43 1.415572e-47 4.199369e-52
#> score.7  3.360880e-33 4.332494e-37 3.636794e-41 2.064602e-45 8.254893e-50
#> score.8  1.669969e-31 2.845360e-35 3.173221e-39 2.404352e-43 1.288387e-47
#> score.9  6.858745e-30 1.536510e-33 2.264684e-37 2.278324e-41 1.627652e-45
#> score.10 2.361974e-28 6.922737e-32 1.341910e-35 1.783643e-39 1.690474e-43
#> score.11 6.901695e-27 2.634066e-30 6.683721e-34 1.168301e-37 1.462078e-41
#> score.12 1.728441e-25 8.551322e-29 2.827685e-32 6.471146e-36 1.064536e-39
#> score.13 3.742256e-24 2.389588e-27 1.025312e-30 3.058830e-34 6.586001e-38
#> score.14 7.057935e-23 5.791976e-26 3.211283e-29 1.243723e-32 3.490360e-36
#> score.15 1.167327e-21 1.225960e-24 8.747223e-28 4.380356e-31 1.595851e-34
#> score.16 1.703252e-20 2.279713e-23 2.084796e-26 1.344580e-29 6.334578e-33
#> score.17 2.204414e-19 3.744395e-22 4.371298e-25 3.616888e-28 2.195203e-31
#> score.18 2.543319e-18 5.458975e-21 8.102695e-24 8.568161e-27 6.674700e-30
#> score.19 2.627987e-17 7.096273e-20 1.333672e-22 1.795425e-25 1.788684e-28
#> score.20 2.442754e-16 8.259909e-19 1.957251e-21 3.341363e-24 4.241629e-27
#> score.21 2.051292e-15 8.643640e-18 2.570900e-20 5.543274e-23 8.933418e-26
#> score.22 1.562772e-14 8.163937e-17 3.033579e-19 8.226209e-22 1.676658e-24
#> score.23 1.084711e-13 6.986857e-16 3.227096e-18 1.095620e-20 2.812983e-23
#> score.24 6.888534e-13 5.439601e-15 3.106113e-17 1.313876e-19 4.231290e-22
#> score.25 4.019709e-12 3.868380e-14 2.715063e-16 1.423331e-18 5.722981e-21
#> score.26 2.164563e-11 2.523485e-13 2.163660e-15 1.397653e-17 6.980641e-20
#> score.27 1.080076e-10 1.516554e-12 1.578463e-14 1.248602e-16 7.702687e-19
#> score.28 5.013390e-10 8.432818e-12 1.058771e-13 1.018847e-15 7.715043e-18
#> score.29 2.172196e-09 4.356514e-11 6.558930e-13 7.626633e-15 7.040973e-17
#> score.30 8.810714e-09 2.098829e-10 3.769123e-12 5.261331e-14 5.880004e-16
#> score.31 3.353159e-08 9.458979e-10 2.017394e-11 3.360731e-13 4.514582e-15
#> score.32 1.199402e-07 3.997439e-09 1.009200e-10 1.996554e-12 3.202665e-14
#> score.33 4.037237e-07 1.586786e-08 4.730676e-10 1.107383e-11 2.109492e-13
#> score.34 1.280069e-06 5.922871e-08 2.081426e-09 5.750711e-11 1.295620e-12
#> score.35 3.826275e-06 2.080426e-07 8.603879e-09 2.800927e-10 7.443830e-12
#> score.36 1.079100e-05 6.881092e-07 3.342950e-08 1.280429e-09 4.007918e-11
#> score.37 2.873609e-05 2.144598e-06 1.221308e-07 5.494544e-09 2.023305e-10
#> score.38 7.230744e-05 6.303073e-06 4.197590e-07 2.213185e-08 9.573500e-10
#> score.39 1.720203e-04 1.748316e-05 1.358243e-06 8.369806e-08 4.243321e-09
#> score.40 3.870770e-04 4.579843e-05 4.141437e-06 2.973855e-07 1.761562e-08
#> score.41 8.240436e-04 1.133607e-04 1.190983e-05 9.937774e-07 6.853142e-08
#> score.42 1.660015e-03 2.652047e-04 3.232474e-05 3.127010e-06 2.501600e-07
#> score.43 3.164813e-03 5.864967e-04 8.283343e-05 9.273797e-06 8.581375e-07
#> score.44 5.711330e-03 1.226176e-03 2.004352e-04 2.593673e-05 2.770166e-06
#> score.45 9.758268e-03 2.423811e-03 4.579846e-04 6.841817e-05 8.422195e-06
#> score.46 1.578860e-02 4.530951e-03 9.882535e-04 1.702189e-04 2.412178e-05
#> score.47 2.419394e-02 8.011665e-03 2.014233e-03 3.994166e-04 6.507257e-05
#> score.48 3.511216e-02 1.340206e-02 3.878752e-03 8.841206e-04 1.653288e-04
#> score.49 4.825157e-02 2.121030e-02 7.058531e-03 1.846781e-03 3.956866e-04
#> score.50 6.276021e-02 3.175224e-02 1.213965e-02 3.641501e-03 8.925197e-04
#> score.51 7.720954e-02 4.494491e-02 1.972885e-02 6.778920e-03 1.898281e-03
#> score.52 8.974549e-02 6.011348e-02 3.028535e-02 1.191214e-02 3.807723e-03
#> score.53 9.841451e-02 7.589210e-02 4.388499e-02 1.975125e-02 7.201927e-03
#> score.54 1.016087e-01 9.030143e-02 5.996673e-02 3.088171e-02 1.283840e-02
#> score.55 9.851031e-02 1.010531e-01 7.715165e-02 4.548673e-02 2.155688e-02
#> score.56 8.938102e-02 1.060595e-01 9.325164e-02 6.301776e-02 3.406285e-02
#> score.57 7.557026e-02 1.040240e-01 1.055718e-01 8.192254e-02 5.057125e-02
#> score.58 5.920949e-02 9.490489e-02 1.115192e-01 9.960746e-02 7.035772e-02
#> score.59 4.268379e-02 8.005103e-02 1.093694e-01 1.128040e-01 9.139237e-02
#> score.60 2.805504e-02 6.192210e-02 9.891681e-02 1.183529e-01 1.103392e-01
#> score.61 1.662298e-02 4.345973e-02 8.173895e-02 1.141911e-01 1.231126e-01
#> score.62 8.757199e-03 2.729958e-02 6.092966e-02 1.002226e-01 1.259095e-01
#> score.63 4.034592e-03 1.508978e-02 4.028544e-02 7.877027e-02 1.165000e-01
#> score.64 1.593611e-03 7.188949e-03 2.312522e-02 5.426234e-02 9.557988e-02
#> score.65 5.266084e-04 2.877517e-03 1.122083e-02 3.185996e-02 6.757821e-02
#> score.66 1.410602e-04 9.366361e-04 4.448194e-03 1.538285e-02 3.964484e-02
#> score.67 2.931893e-05 2.371043e-04 1.375959e-03 5.822348e-03 1.834826e-02
#> score.68 4.422290e-06 4.362478e-05 3.100486e-04 1.610150e-03 6.229149e-03
#> score.69 4.294291e-07 5.172520e-06 4.508672e-05 2.878687e-04 1.370081e-03
#> score.70 2.010177e-08 2.958182e-07 3.164990e-06 2.486524e-05 1.457122e-04
#>              theta.31     theta.32     theta.33     theta.34     theta.35
#> score.0  1.269793e-74 2.812760e-80 4.874417e-86 6.889136e-92 8.232569e-98
#> score.1  5.115848e-71 1.644423e-76 4.162279e-82 8.645205e-88 1.527112e-93
#> score.2  8.773085e-68 4.005001e-73 1.445254e-78 4.294153e-84 1.088295e-89
#> score.3  8.704960e-65 5.566939e-70 2.823426e-75 1.182504e-80 4.235658e-86
#> score.4  5.693751e-62 5.049473e-67 3.562670e-72 2.081777e-77 1.043185e-82
#> score.5  2.642589e-59 3.223546e-64 3.138306e-69 2.537850e-74 1.764852e-79
#> score.6  9.128734e-57 1.521068e-61 2.029196e-66 2.255248e-71 2.161433e-76
#> score.7  2.427668e-54 5.491196e-59 9.975889e-64 1.514256e-68 1.987476e-73
#> score.8  5.097030e-52 1.556175e-56 3.827794e-61 7.889138e-66 1.409591e-70
#> score.9  8.616541e-50 3.531963e-54 1.169897e-58 3.255700e-63 7.873855e-68
#> score.10 1.191552e-47 6.524290e-52 2.895025e-56 1.082064e-60 3.522875e-65
#> score.11 1.365655e-45 9.940123e-50 5.879529e-54 2.936534e-58 1.280297e-62
#> score.12 1.311647e-43 1.263224e-47 9.912919e-52 6.583830e-56 3.825015e-60
#> score.13 1.065774e-41 1.352105e-45 1.401317e-49 1.231968e-53 9.493030e-58
#> score.14 7.387162e-40 1.229307e-43 1.675424e-47 1.941277e-51 1.975355e-55
#> score.15 4.399623e-38 9.564848e-42 1.707315e-45 2.596593e-49 3.474901e-53
#> score.16 2.266102e-36 6.411345e-40 1.493093e-43 2.969219e-47 5.206096e-51
#> score.17 1.015225e-34 3.724314e-38 1.127483e-41 2.921276e-45 6.687086e-49
#> score.18 3.976389e-33 1.884804e-36 7.392057e-40 2.486975e-43 7.407883e-47
#> score.19 1.367878e-31 8.349465e-35 4.228406e-38 1.841394e-41 7.115079e-45
#> score.20 4.149711e-30 3.251259e-33 2.119502e-36 1.191128e-39 5.952895e-43
#> score.21 1.114290e-28 1.117053e-31 9.345713e-35 6.758309e-38 4.356440e-41
#> score.22 2.657228e-27 3.397635e-30 3.637419e-33 3.375264e-36 2.798712e-39
#> score.23 5.644425e-26 9.176083e-29 1.253381e-31 1.488325e-34 1.583320e-37
#> score.24 1.070974e-24 2.206388e-27 3.833838e-30 5.809842e-33 7.909236e-36
#> score.25 1.819899e-23 4.734927e-26 1.043436e-28 2.012379e-31 3.496666e-34
#> score.26 2.776748e-22 9.089571e-25 2.532171e-27 6.197291e-30 1.370788e-32
#> score.27 3.813974e-21 1.564341e-23 5.489762e-26 1.699803e-28 4.772980e-31
#> score.28 4.729098e-20 2.419102e-22 1.065242e-24 4.158882e-27 1.478099e-29
#> score.29 5.309876e-19 3.369495e-21 1.853489e-23 9.090109e-26 4.075839e-28
#> score.30 5.418096e-18 4.239039e-20 2.897804e-22 1.777527e-24 1.001810e-26
#> score.31 5.045114e-17 4.832777e-19 4.080685e-21 3.114817e-23 2.197172e-25
#> score.32 4.307556e-16 5.012956e-18 5.191422e-20 4.901161e-22 4.305092e-24
#> score.33 3.390045e-15 4.753874e-17 5.989502e-19 6.943477e-21 7.548217e-23
#> score.34 2.472376e-14 4.144354e-16 6.297126e-18 8.888779e-20 1.187037e-21
#> score.35 1.679015e-13 3.340901e-15 6.068311e-17 1.033243e-18 1.680105e-20
#> score.36 1.065609e-12 2.504032e-14 5.394646e-16 1.097329e-17 2.150825e-19
#> score.37 6.332620e-12 1.752219e-13 4.451715e-15 1.072381e-16 2.507014e-18
#> score.38 3.524596e-11 1.147148e-12 3.426435e-14 9.712619e-16 2.682190e-17
#> score.39 1.835546e-10 7.025615e-12 2.465464e-13 8.198057e-15 2.655807e-16
#> score.40 8.933227e-10 4.018295e-11 1.657469e-12 6.464390e-14 2.449432e-15
#> score.41 4.060089e-09 2.141691e-10 1.038261e-11 4.755867e-13 2.109325e-14
#> score.42 1.723883e-08 1.062228e-09 6.040027e-11 3.251235e-12 1.691904e-13
#> score.43 6.847927e-08 4.903171e-09 3.255702e-10 2.055375e-11 1.256521e-12
#> score.44 2.550317e-07 2.110091e-08 1.625770e-09 1.197585e-10 8.582704e-12
#> score.45 8.922180e-07 8.490147e-08 7.538032e-09 6.429181e-10 5.368046e-11
#> score.46 2.935777e-06 3.202872e-07 3.257810e-08 3.189891e-09 3.073795e-10
#> score.47 9.088150e-06 1.134858e-06 1.317605e-07 1.470725e-08 1.618757e-09
#> score.48 2.646038e-05 3.778017e-06 4.999109e-07 6.336139e-08 7.901534e-09
#> score.49 7.243842e-05 1.180969e-05 1.779683e-06 2.558635e-07 3.601889e-08
#> score.50 1.865083e-04 3.464419e-05 5.937585e-06 9.681655e-07 1.538672e-07
#> score.51 4.519579e-04 9.540730e-05 1.854743e-05 3.424979e-06 6.149028e-07
#> score.52 1.031562e-03 2.469408e-04 5.427783e-05 1.131154e-05 2.289678e-06
#> score.53 2.218250e-03 6.014099e-04 1.490970e-04 3.492459e-05 7.931919e-06
#> score.54 4.492650e-03 1.378630e-03 3.851109e-04 1.011432e-04 2.564658e-05
#> score.55 8.564637e-03 2.972618e-03 9.354442e-04 2.754045e-04 7.782745e-05
#> score.56 1.535963e-02 6.023718e-03 2.133809e-03 7.045027e-04 2.221818e-04
#> score.57 2.589390e-02 1.146687e-02 4.565249e-03 1.688284e-03 5.947116e-04
#> score.58 4.096900e-02 2.049851e-02 9.162480e-03 3.785311e-03 1.485402e-03
#> score.59 6.064782e-02 3.435192e-02 1.725704e-02 7.954713e-03 3.462815e-03
#> score.60 8.363534e-02 5.375242e-02 3.043424e-02 1.569018e-02 7.574667e-03
#> score.61 1.069123e-01 7.811106e-02 4.996269e-02 2.892656e-02 1.556861e-02
#> score.62 1.259475e-01 1.048691e-01 7.583218e-02 4.937563e-02 2.977028e-02
#> score.63 1.354840e-01 1.293658e-01 1.059507e-01 7.745179e-02 5.220792e-02
#> score.64 1.308983e-01 1.450747e-01 1.357080e-01 1.116016e-01 8.368133e-02
#> score.65 1.105228e-01 1.445358e-01 1.570746e-01 1.473571e-01 1.236766e-01
#> score.66 7.838230e-02 1.229348e-01 1.584561e-01 1.737518e-01 1.673666e-01
#> score.67 4.424219e-02 8.424981e-02 1.310210e-01 1.718994e-01 1.959925e-01
#> score.68 1.841595e-02 4.289372e-02 8.131675e-02 1.295226e-01 1.784221e-01
#> score.69 4.979583e-03 1.423562e-02 3.305328e-02 6.432325e-02 1.079739e-01
#> score.70 6.516656e-04 2.289315e-03 6.520906e-03 1.553972e-02 3.188588e-02
#>               theta.36      theta.37      theta.38      theta.39      theta.40
#> score.0  8.574526e-104 7.980937e-110 6.775500e-116 5.334540e-122 3.948221e-128
#> score.1   2.363995e-99 3.287186e-105 4.189275e-111 4.973690e-117 5.574321e-123
#> score.2   2.423384e-95 4.858352e-101 8.944688e-107 1.536884e-112 2.496797e-118
#> score.3   1.336307e-91  3.804220e-97 9.966777e-103 2.441817e-108 5.667169e-114
#> score.4   4.615965e-88  1.847518e-93  6.821026e-99 2.360184e-104 7.752985e-110
#> score.5   1.086584e-84  6.066343e-90  3.131556e-95 1.518515e-100 7.005737e-106
#> score.6   1.838809e-81  1.422040e-86  1.019226e-91  6.877236e-97 4.424261e-102
#> score.7   2.321400e-78  2.470605e-83  2.442289e-88  2.277563e-93  2.028912e-98
#> score.8   2.246670e-75  3.269987e-80  4.429707e-85  5.671503e-90  6.948566e-95
#> score.9   1.702466e-72  3.368281e-77  6.213870e-82  1.085288e-86  1.816680e-91
#> score.10  1.027507e-69  2.747374e-74  6.861296e-79  1.624770e-83  3.692714e-88
#> score.11  5.010329e-67  1.800603e-71  6.053490e-76  1.932487e-80  5.928964e-85
#> score.12  1.998308e-64  9.602946e-69  4.323516e-73  1.850979e-77  7.625967e-82
#> score.13  6.589596e-62  4.214329e-66  2.528949e-70  1.445106e-74  7.957605e-79
#> score.14  1.813930e-59  1.537163e-63  1.224131e-67  9.296586e-72  6.813453e-76
#> score.15  4.204141e-57  4.701786e-61  4.949370e-65  4.976194e-69  4.835639e-73
#> score.16  8.267540e-55  1.215748e-58  1.685529e-62  2.235608e-66  2.870514e-70
#> score.17  1.389051e-52  2.676602e-56  4.871116e-60  8.495251e-64  1.436648e-67
#> score.18  2.006247e-50  5.049717e-54  1.202565e-57  2.749230e-61  6.104884e-65
#> score.19  2.504721e-48  8.210386e-52  2.551086e-55  7.622849e-59  2.216222e-62
#> score.20  2.716136e-46  1.156249e-49  4.674314e-53  1.820478e-56  6.910130e-60
#> score.21  2.569288e-44  1.416570e-47  7.430902e-51  3.761894e-54  1.859136e-57
#> score.22  2.127931e-42  1.515581e-45  1.028932e-48  6.752990e-52  4.333288e-55
#> score.23  1.548029e-40  1.420687e-43  1.245088e-46  1.056613e-49  8.779566e-53
#> score.24  9.918994e-39  1.170046e-41  1.320388e-44  1.445083e-47  1.550617e-50
#> score.25  5.610797e-37  8.485924e-40  1.229989e-42  1.731551e-45  2.392846e-48
#> score.26  2.807239e-35  5.430104e-38  1.008351e-40  1.821158e-43  3.232198e-46
#> score.27  1.244242e-33  3.070286e-36  7.285569e-39  1.683617e-41  3.826942e-44
#> score.28  4.891411e-32  1.535686e-34  4.644245e-37  1.369477e-39  3.975468e-42
#> score.29  1.707190e-30  6.800299e-33  2.613770e-35  9.807222e-38  3.625286e-40
#> score.30  5.293822e-29  2.667359e-31  1.299200e-33  6.184739e-36  2.902562e-38
#> score.31  1.459359e-27  9.270232e-30  5.703885e-32  3.434311e-34  2.039926e-36
#> score.32  3.578670e-26  2.855104e-28  2.211456e-30  1.678531e-32  1.257771e-34
#> score.33  7.812382e-25  7.793555e-27  7.569396e-29  7.216297e-31  6.797834e-33
#> score.34  1.520095e-23  1.886124e-25  2.286524e-27  2.726789e-29  3.216929e-31
#> score.35  2.641539e-22  4.050177e-24  6.094921e-26  9.048915e-28  1.331276e-29
#> score.36  4.113194e-21  7.730380e-23  1.434262e-24  2.635815e-26  4.811855e-28
#> score.37  5.768623e-20  1.315770e-21  2.984220e-23  6.740116e-25  1.517668e-26
#> score.38  7.340548e-19  2.008325e-20  5.509112e-22  1.515353e-23  4.176739e-25
#> score.39  8.554154e-18  2.772167e-19  9.081259e-21  3.007245e-22  1.004706e-23
#> score.40  9.217865e-17  3.498584e-18  1.350145e-19  5.308546e-21  2.122791e-22
#> score.41  9.252564e-16  4.083386e-17  1.834556e-18  8.439791e-20  3.978482e-21
#> score.42  8.668197e-15  4.443006e-16  2.308974e-17  1.228244e-18  6.720143e-20
#> score.43  7.546529e-14  4.509703e-15  2.713429e-16  1.661325e-17  1.043579e-18
#> score.44  6.053302e-13  4.238531e-14  2.970746e-15  2.101935e-16  1.514001e-17
#> score.45  4.434232e-12  3.644816e-13  2.993921e-14  2.468988e-15  2.055626e-16
#> score.46  2.950702e-11  2.836065e-12  2.733398e-13  2.643956e-14  2.571073e-15
#> score.47  1.784935e-10  1.985902e-11  2.232021e-12  2.530429e-13  2.888755e-14
#> score.48  9.885768e-10  1.255047e-10  1.623225e-11  2.136284e-12  2.851014e-13
#> score.49  5.069047e-09  7.240571e-10  1.058977e-10  1.589910e-11  2.445077e-12
#> score.50  2.429862e-08  3.873518e-09  6.310199e-10  1.058744e-10  1.834551e-11
#> score.51  1.091817e-07  1.943182e-08  3.507000e-09  6.483802e-10  1.237614e-10
#> score.52  4.576823e-07  9.134799e-08  1.835663e-08  3.742603e-09  7.802669e-10
#> score.53  1.778860e-06  3.984401e-07  8.971059e-08  2.038544e-08  4.692112e-09
#> score.54  6.408429e-06  1.600459e-06  4.027720e-07  1.025004e-07  2.639949e-08
#> score.55  2.155369e-05  5.946357e-06  1.653685e-06  4.668458e-07  1.341369e-07
#> score.56  6.822514e-05  2.071801e-05  6.303673e-06  1.942109e-06  6.104091e-07
#> score.57  2.032087e-04  6.825020e-05  2.277089e-05  7.618593e-06  2.578897e-06
#> score.58  5.647685e-04  2.108188e-04  7.791265e-05  2.867400e-05  1.056110e-05
#> score.59  1.456483e-03  6.017178e-04  2.467016e-04  1.009261e-04  4.129564e-05
#> score.60  3.506274e-03  1.586826e-03  7.126354e-04  3.207322e-04  1.454046e-04
#> score.61  7.959626e-03  3.937235e-03  1.914801e-03  9.282893e-04  4.535238e-04
#> score.62  1.696635e-02  9.280826e-03  4.933767e-03  2.577570e-03  1.337510e-03
#> score.63  3.323583e-02  2.030592e-02  1.203064e-02  6.960727e-03  3.954445e-03
#> score.64  5.895074e-02  3.989159e-02  2.630898e-02  1.705115e-02  1.090551e-02
#> score.65  9.590624e-02  7.064808e-02  5.055065e-02  3.570223e-02  2.513393e-02
#> score.66  1.457892e-01  1.178703e-01  9.054135e-02  6.746466e-02  4.964267e-02
#> score.67  1.993013e-01  1.848254e-01  1.593089e-01  1.297136e-01  1.011692e-01
#> score.68  2.179849e-01  2.413007e-01  2.463829e-01  2.355264e-01  2.134053e-01
#> score.69  1.603081e-01  2.150385e-01  2.652911e-01  3.054776e-01  3.323383e-01
#> score.70  5.776839e-02  9.440499e-02  1.416684e-01  1.981373e-01  2.614590e-01
#>               theta.41
#> score.0  2.777462e-134
#> score.1  5.961347e-129
#> score.2  3.876068e-124
#> score.3  1.259156e-119
#> score.4  2.443166e-115
#> score.5  3.107126e-111
#> score.6  2.741552e-107
#> score.7  1.744082e-103
#> score.8  8.228165e-100
#> score.9   2.943409e-96
#> score.10  8.134205e-93
#> score.11  1.765269e-89
#> score.12  3.052926e-86
#> score.13  4.263590e-83
#> score.14  4.865631e-80
#> score.15  4.585569e-77
#> score.16  3.602399e-74
#> score.17  2.378476e-71
#> score.18  1.329323e-68
#> score.19  6.328496e-66
#> score.20  2.580243e-63
#> score.21  9.051832e-61
#> score.22  2.743218e-58
#> score.23  7.206043e-56
#> score.24  1.645381e-53
#> score.25  3.273128e-51
#> score.26  5.682891e-49
#> score.27  8.623222e-47
#> score.28  1.144614e-44
#> score.29  1.329696e-42
#> score.30  1.352044e-40
#> score.31  1.202955e-38
#> score.32  9.359274e-37
#> score.33  6.361113e-35
#> score.34  3.771810e-33
#> score.35  1.948091e-31
#> score.36  8.749109e-30
#> score.37  3.411069e-28
#> score.38  1.153068e-26
#> score.39  3.379213e-25
#> score.40  8.605125e-24
#> score.41  1.916600e-22
#> score.42  3.783361e-21
#> score.43  6.757910e-20
#> score.44  1.118396e-18
#> score.45  1.738962e-17
#> score.46  2.521206e-16
#> score.47  3.318725e-15
#> score.48  3.844657e-14
#> score.49  3.834236e-13
#> score.50  3.276943e-12
#> score.51  2.449137e-11
#> score.52  1.676048e-10
#> score.53  1.098829e-09
#> score.54  6.881066e-09
#> score.55  3.920336e-08
#> score.56  1.964651e-07
#> score.57  8.901941e-07
#> score.58  3.912986e-06
#> score.59  1.691493e-05
#> score.60  6.649623e-05
#> score.61  2.249240e-04
#> score.62  6.963588e-04
#> score.63  2.216669e-03
#> score.64  6.897122e-03
#> score.65  1.771751e-02
#> score.66  3.659136e-02
#> score.67  7.650117e-02
#> score.68  1.851640e-01
#> score.69  3.450608e-01
#> score.70  3.288416e-01