Polytomous Response Model (PRM) Probabilities (GRM and GPCM)

This function computes the probability of selecting each response category for an item, given a set of theta values, using the graded response model (GRM) or the (generalized) partial credit model (GPCM).

Usage

prm(theta, a, d, D = 1, pr.model = c("GRM", "GPCM"))

Arguments

theta

A numeric vector of ability values (latent traits).

a

A numeric vector of item discrimination (slope) parameters.

d

A numeric vector of item difficulty (or threshold) parameters.

D

A scaling constant used in IRT models to make the logistic function closely approximate the normal ogive function. A value of 1.7 is commonly used for this purpose. Default is 1.

pr.model

A character string specifying the polytomous IRT model. Available options are "GRM" for the graded response model and "GPCM" for the (generalized) partial credit model.

Value

A matrix of category response probabilities, where each row corresponds to a theta value and each column represents a score category of the item.

Details

When computing category probabilities using the partial credit model (PCM), set a = 1.

For pr.model = "GPCM", the vector d should contain threshold parameters that define the boundaries between adjacent score categories. In the irtQ package, these thresholds are expressed as the item location (overall difficulty) minus the step parameters for each category. If an item has K score categories, K - 1 threshold parameters must be provided; the first is assumed to be 0. For example, for a GPCM item with five categories, provide four threshold parameters.

For more details on the parameterization of the (generalized) partial credit model, refer to the IRT Models section in the irtQ-package documentation.

See also

Author

Hwanggyu Lim hglim83@gmail.com

Examples

## Category probabilities for an item with four categories
## using the generalized partial credit model (GPCM)
prm(theta = c(-0.2, 0, 0.5), a = 1.4, d = c(-0.2, 0, 0.5), D = 1, pr.model = "GPCM")
#>            [,1]      [,2]      [,3]       [,4]
#> [1,] 0.32900822 0.3290082 0.2486591 0.09332451
#> [2,] 0.23237945 0.3074682 0.3074682 0.15268418
#> [3,] 0.06946582 0.1850886 0.3727228 0.37272276

## Category probabilities for an item with five categories
## using the graded response model (GRM)
prm(theta = c(-0.2, 0, 0.5), a = 1.2, d = c(-0.4, -0.2, 0.4, 1.5), D = 1, pr.model = "GRM")
#>           [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.4402864 0.05971365 0.1726070 0.2123263 0.1150667
#> [2,] 0.3822521 0.05803423 0.1774615 0.2404011 0.1418511
#> [3,] 0.2535060 0.04802877 0.1685012 0.2984888 0.2314752