Compute Item/Test Characteristic Functions

This function computes item category probabilities, item characteristic curves (ICCs), and the test characteristic curve (TCC) for a given set of theta values. The returned object can be used to visualize these functions using plot.traceline().

Usage

traceline(x, ...)

# Default S3 method
traceline(x, theta, D = 1, ...)

# S3 method for class 'est_item'
traceline(x, theta, ...)

# S3 method for class 'est_irt'
traceline(x, theta, ...)

Arguments

x

A data frame containing item metadata (e.g., item parameters, number of categories, IRT model types, etc.); or an object of class est_irt obtained from est_irt(), or est_item from est_item().

See est_irt() or simdat() for more details about the item metadata. This data frame can be easily created using the shape_df() function.

...

Further arguments passed to or from other methods.

theta

A numeric vector of theta values at which item and test characteristic curves are computed.

D

A scaling constant used in IRT models to make the logistic function closely approximate the normal ogive function. A value of 1.7 is commonly used for this purpose. Default is 1.

Value

This function returns an object of class traceline, which is a list containing the following components:

prob.cats

A list of data frames containing the category response probabilities for each item across the specified theta values. Each data frame corresponds to an item, with rows representing theta values and columns representing response categories (e.g., "resp.0", "resp.1", ...).

icc

A numeric matrix representing ICCs. Each column corresponds to an item, and each row represents the expected item score at a given theta value. The column names are the item IDs.

tcc

A numeric vector representing the TCC, computed as the sum of expected item scores across all items at each theta value.

theta

A numeric vector of theta values at which the item and test information functions are evaluated. This matches the user-supplied theta argument.

Details

This function computes the item and test characteristic functions commonly used in IRT. For each item, the function computes the category response probabilities across a specified set of theta values. These probabilities are used to derive:

The item characteristic curve (ICC), which represents the expected score of each item as a function of theta.

The test characteristic curve (TCC), which is the sum of expected item scores at each theta value.

The output object can be visualized using the plot.traceline to inspect the relationship between ability levels (theta) and expected item/test scores.

If the input x is an object of class est_item or est_irt, the function automatically extracts item parameter estimates and the scaling constant D from the object. Otherwise, a properly formatted item metadata data frame must be provided.

Methods (by class)

traceline(default): Default method to compute the item category probabilities, item characteristic function, and test characteristic function for a data frame x containing the item metadata.

traceline(est_item): An object created by the function est_item().

traceline(est_irt): An object created by the function est_irt().

See also

Author

Hwanggyu Lim hglim83@gmail.com

Examples

## Example using a "-prm.txt" file exported from flexMIRT

# Import the "-prm.txt" output file from flexMIRT
flex_prm <- system.file("extdata", "flexmirt_sample-prm.txt", package = "irtQ")

# Read the item parameters and convert them into item metadata
test_flex <- bring.flexmirt(file = flex_prm, "par")$Group1$full_df

# Define a sequence of theta values
theta <- seq(-3, 3, 0.5)

# Compute item category probabilities, ICCs,
# and the TCC for the given theta values
traceline(x = test_flex, theta, D = 1)
#> $prob.cats
#> $prob.cats$CMC1
#>          resp.0    resp.1
#>  [1,] 0.7147254 0.2852746
#>  [2,] 0.7042014 0.2957986
#>  [3,] 0.6893382 0.3106618
#>  [4,] 0.6686717 0.3313283
#>  [5,] 0.6405512 0.3594488
#>  [6,] 0.6033936 0.3966064
#>  [7,] 0.5561521 0.4438479
#>  [8,] 0.4989494 0.5010506
#>  [9,] 0.4336384 0.5663616
#> [10,] 0.3638896 0.6361104
#> [11,] 0.2945212 0.7054788
#> [12,] 0.2302486 0.7697514
#> [13,] 0.1744896 0.8255104
#> 
#> $prob.cats$CMC2
#>             resp.0    resp.1
#>  [1,] 0.8056920416 0.1943080
#>  [2,] 0.7767979880 0.2232020
#>  [3,] 0.7102335670 0.2897664
#>  [4,] 0.5802836604 0.4197163
#>  [5,] 0.3925707104 0.6074293
#>  [6,] 0.2127308557 0.7872691
#>  [7,] 0.0968187150 0.9031813
#>  [8,] 0.0399422537 0.9600577
#>  [9,] 0.0157548536 0.9842451
#> [10,] 0.0061004964 0.9938995
#> [11,] 0.0023450464 0.9976550
#> [12,] 0.0008989032 0.9991011
#> [13,] 0.0003441943 0.9996558
#> 
#> $prob.cats$CMC3
#>           resp.0    resp.1
#>  [1,] 0.86311101 0.1368890
#>  [2,] 0.84262217 0.1573778
#>  [3,] 0.81198693 0.1880131
#>  [4,] 0.76762994 0.2323701
#>  [5,] 0.70630633 0.2936937
#>  [6,] 0.62673098 0.3732690
#>  [7,] 0.53154979 0.4684502
#>  [8,] 0.42819451 0.5718055
#>  [9,] 0.32710652 0.6728935
#> [10,] 0.23785847 0.7621415
#> [11,] 0.16591206 0.8340879
#> [12,] 0.11204656 0.8879534
#> [13,] 0.07390907 0.9260909
#> 
#> $prob.cats$CMC4
#>           resp.0    resp.1
#>  [1,] 0.74987378 0.2501262
#>  [2,] 0.71934988 0.2806501
#>  [3,] 0.67298037 0.3270196
#>  [4,] 0.60677430 0.3932257
#>  [5,] 0.52015640 0.4798436
#>  [6,] 0.41892879 0.5810712
#>  [7,] 0.31512267 0.6848773
#>  [8,] 0.22200626 0.7779937
#>  [9,] 0.14798658 0.8520134
#> [10,] 0.09459848 0.9054015
#> [11,] 0.05873270 0.9412673
#> [12,] 0.03577417 0.9642258
#> [13,] 0.02152903 0.9784710
#> 
#> $prob.cats$CMC5
#>           resp.0    resp.1
#>  [1,] 0.77535126 0.2246487
#>  [2,] 0.74447519 0.2555248
#>  [3,] 0.70140146 0.2985985
#>  [4,] 0.64394568 0.3560543
#>  [5,] 0.57172279 0.4282772
#>  [6,] 0.48741847 0.5125815
#>  [7,] 0.39711526 0.6028847
#>  [8,] 0.30886877 0.6911312
#>  [9,] 0.23003706 0.7699629
#> [10,] 0.16506741 0.8349326
#> [11,] 0.11498573 0.8850143
#> [12,] 0.07833425 0.9216658
#> [13,] 0.05251860 0.9474814
#> 
#> $prob.cats$CMC6
#>           resp.0     resp.1
#>  [1,] 0.92585317 0.07414683
#>  [2,] 0.92322624 0.07677376
#>  [3,] 0.91715122 0.08284878
#>  [4,] 0.90327539 0.09672461
#>  [5,] 0.87246033 0.12753967
#>  [6,] 0.80810152 0.19189848
#>  [7,] 0.68938214 0.31061786
#>  [8,] 0.51332688 0.48667312
#>  [9,] 0.32159411 0.67840589
#> [10,] 0.17179446 0.82820554
#> [11,] 0.08229944 0.91770056
#> [12,] 0.03713614 0.96286386
#> [13,] 0.01627982 0.98372018
#> 
#> $prob.cats$CMC7
#>          resp.0    resp.1
#>  [1,] 0.8579915 0.1420085
#>  [2,] 0.8455802 0.1544198
#>  [3,] 0.8267661 0.1732339
#>  [4,] 0.7988060 0.2011940
#>  [5,] 0.7584552 0.2415448
#>  [6,] 0.7026211 0.2973789
#>  [7,] 0.6296975 0.3703025
#>  [8,] 0.5413154 0.4586846
#>  [9,] 0.4434103 0.5565897
#> [10,] 0.3451938 0.6548062
#> [11,] 0.2559922 0.7440078
#> [12,] 0.1820046 0.8179954
#> [13,] 0.1251172 0.8748828
#> 
#> $prob.cats$CMC8
#>          resp.0    resp.1
#>  [1,] 0.8511265 0.1488735
#>  [2,] 0.8341106 0.1658894
#>  [3,] 0.8094291 0.1905709
#>  [4,] 0.7744756 0.2255244
#>  [5,] 0.7266181 0.2733819
#>  [6,] 0.6640336 0.3359664
#>  [7,] 0.5869292 0.4130708
#>  [8,] 0.4986224 0.5013776
#>  [9,] 0.4055560 0.5944440
#> [10,] 0.3156869 0.6843131
#> [11,] 0.2359426 0.7640574
#> [12,] 0.1703212 0.8296788
#> [13,] 0.1195907 0.8804093
#> 
#> $prob.cats$CMC9
#>          resp.0    resp.1
#>  [1,] 0.7180111 0.2819889
#>  [2,] 0.7045350 0.2954650
#>  [3,] 0.6848342 0.3151658
#>  [4,] 0.6566888 0.3433112
#>  [5,] 0.6177738 0.3822262
#>  [6,] 0.5663344 0.4336656
#>  [7,] 0.5022368 0.4977632
#>  [8,] 0.4279914 0.5720086
#>  [9,] 0.3489299 0.6510701
#> [10,] 0.2719352 0.7280648
#> [11,] 0.2032079 0.7967921
#> [12,] 0.1464695 0.8535305
#> [13,] 0.1025726 0.8974274
#> 
#> $prob.cats$CMC10
#>            resp.0    resp.1
#>  [1,] 0.857083405 0.1429166
#>  [2,] 0.848574681 0.1514253
#>  [3,] 0.830831671 0.1691683
#>  [4,] 0.795070184 0.2049298
#>  [5,] 0.727691476 0.2723085
#>  [6,] 0.615487264 0.3845127
#>  [7,] 0.462184463 0.5378155
#>  [8,] 0.300940242 0.6990598
#>  [9,] 0.171918262 0.8280817
#> [10,] 0.089435198 0.9105648
#> [11,] 0.044011535 0.9559885
#> [12,] 0.021032357 0.9789676
#> [13,] 0.009906166 0.9900938
#> 
#> $prob.cats$CMC11
#>           resp.0    resp.1
#>  [1,] 0.80436703 0.1956330
#>  [2,] 0.76797210 0.2320279
#>  [3,] 0.71461314 0.2853869
#>  [4,] 0.64110419 0.3588958
#>  [5,] 0.54806773 0.4519323
#>  [6,] 0.44217454 0.5578255
#>  [7,] 0.33527335 0.6647267
#>  [8,] 0.23965755 0.7603424
#>  [9,] 0.16296242 0.8370376
#> [10,] 0.10664295 0.8933571
#> [11,] 0.06791148 0.9320885
#> [12,] 0.04246188 0.9575381
#> [13,] 0.02623654 0.9737635
#> 
#> $prob.cats$CMC12
#>          resp.0    resp.1
#>  [1,] 0.8897161 0.1102839
#>  [2,] 0.8777271 0.1222729
#>  [3,] 0.8597526 0.1402474
#>  [4,] 0.8332775 0.1667225
#>  [5,] 0.7952817 0.2047183
#>  [6,] 0.7427369 0.2572631
#>  [7,] 0.6736791 0.3263209
#>  [8,] 0.5887498 0.4112502
#>  [9,] 0.4924330 0.5075670
#> [10,] 0.3927470 0.6072530
#> [11,] 0.2988766 0.7011234
#> [12,] 0.2180272 0.7819728
#> [13,] 0.1535732 0.8464268
#> 
#> $prob.cats$CMC13
#>           resp.0    resp.1
#>  [1,] 0.81738997 0.1826100
#>  [2,] 0.81596461 0.1840354
#>  [3,] 0.81302753 0.1869725
#>  [4,] 0.80700968 0.1929903
#>  [5,] 0.79482196 0.2051780
#>  [6,] 0.77070894 0.2292911
#>  [7,] 0.72513489 0.2748651
#>  [8,] 0.64599855 0.3540015
#>  [9,] 0.52687851 0.4731215
#> [10,] 0.38124178 0.6187582
#> [11,] 0.24242234 0.7575777
#> [12,] 0.13818438 0.8618156
#> [13,] 0.07308327 0.9269167
#> 
#> $prob.cats$CMC14
#>           resp.0    resp.1
#>  [1,] 0.74593001 0.2540700
#>  [2,] 0.73924583 0.2607542
#>  [3,] 0.72538651 0.2746135
#>  [4,] 0.69750459 0.3024954
#>  [5,] 0.64467062 0.3553294
#>  [6,] 0.55502997 0.4449700
#>  [7,] 0.42810665 0.5718934
#>  [8,] 0.28778700 0.7122130
#>  [9,] 0.16940028 0.8305997
#> [10,] 0.09024525 0.9097547
#> [11,] 0.04520644 0.9547936
#> [12,] 0.02190122 0.9780988
#> [13,] 0.01043314 0.9895669
#> 
#> $prob.cats$CMC15
#>           resp.0    resp.1
#>  [1,] 0.86428921 0.1357108
#>  [2,] 0.84367116 0.1563288
#>  [3,] 0.80686373 0.1931363
#>  [4,] 0.74473259 0.2552674
#>  [5,] 0.64914575 0.3508543
#>  [6,] 0.52121820 0.4787818
#>  [7,] 0.37855842 0.6214416
#>  [8,] 0.24849673 0.7515033
#>  [9,] 0.14996815 0.8500319
#> [10,] 0.08529603 0.9147040
#> [11,] 0.04674675 0.9532533
#> [12,] 0.02507550 0.9749245
#> [13,] 0.01329207 0.9867079
#> 
#> $prob.cats$CMC16
#>            resp.0     resp.1
#>  [1,] 0.946682295 0.05331770
#>  [2,] 0.942385134 0.05761487
#>  [3,] 0.930629567 0.06937043
#>  [4,] 0.899407544 0.10059246
#>  [5,] 0.822606734 0.17739327
#>  [6,] 0.664663669 0.33533633
#>  [7,] 0.433198865 0.56680113
#>  [8,] 0.219997466 0.78000253
#>  [9,] 0.092841809 0.90715819
#> [10,] 0.035592722 0.96440728
#> [11,] 0.013104888 0.98689511
#> [12,] 0.004751194 0.99524881
#> [13,] 0.001712807 0.99828719
#> 
#> $prob.cats$CMC17
#>           resp.0    resp.1
#>  [1,] 0.80678294 0.1932171
#>  [2,] 0.79244762 0.2075524
#>  [3,] 0.76510111 0.2348989
#>  [4,] 0.71543778 0.2845622
#>  [5,] 0.63281054 0.3671895
#>  [6,] 0.51351214 0.4864879
#>  [7,] 0.37232016 0.6276798
#>  [8,] 0.23972942 0.7602706
#>  [9,] 0.13967345 0.8603266
#> [10,] 0.07592744 0.9240726
#> [11,] 0.03958551 0.9604145
#> [12,] 0.02016775 0.9798322
#> [13,] 0.01015126 0.9898487
#> 
#> $prob.cats$CMC18
#>           resp.0    resp.1
#>  [1,] 0.72635186 0.2736481
#>  [2,] 0.72563069 0.2743693
#>  [3,] 0.72395258 0.2760474
#>  [4,] 0.72006487 0.2799351
#>  [5,] 0.71114901 0.2888510
#>  [6,] 0.69116909 0.3088309
#>  [7,] 0.64862447 0.3513755
#>  [8,] 0.56712459 0.4328754
#>  [9,] 0.43849373 0.5615063
#> [10,] 0.28668658 0.7133134
#> [11,] 0.15854345 0.8414565
#> [12,] 0.07758302 0.9224170
#> [13,] 0.03539077 0.9646092
#> 
#> $prob.cats$CMC19
#>             resp.0    resp.1
#>  [1,] 8.121816e-01 0.1878184
#>  [2,] 7.949277e-01 0.2050723
#>  [3,] 7.446709e-01 0.2553291
#>  [4,] 6.201218e-01 0.3798782
#>  [5,] 4.049536e-01 0.5950464
#>  [6,] 1.926246e-01 0.8073754
#>  [7,] 7.226002e-02 0.9277400
#>  [8,] 2.420686e-02 0.9757931
#>  [9,] 7.777016e-03 0.9922230
#> [10,] 2.464035e-03 0.9975360
#> [11,] 7.772219e-04 0.9992228
#> [12,] 2.448108e-04 0.9997552
#> [13,] 7.707664e-05 0.9999229
#> 
#> $prob.cats$CMC20
#>             resp.0    resp.1
#>  [1,] 0.7084807107 0.2915193
#>  [2,] 0.6255844047 0.3744156
#>  [3,] 0.5039115989 0.4960884
#>  [4,] 0.3595862286 0.6404138
#>  [5,] 0.2260063610 0.7739936
#>  [6,] 0.1279328078 0.8720672
#>  [7,] 0.0674930789 0.9325069
#>  [8,] 0.0341749546 0.9658250
#>  [9,] 0.0169290514 0.9830709
#> [10,] 0.0082929370 0.9917071
#> [11,] 0.0040399513 0.9959600
#> [12,] 0.0019627408 0.9980373
#> [13,] 0.0009523008 0.9990477
#> 
#> $prob.cats$CMC21
#>             resp.0    resp.1
#>  [1,] 0.8339733016 0.1660267
#>  [2,] 0.7844499663 0.2155500
#>  [3,] 0.6917362496 0.3082638
#>  [4,] 0.5460663354 0.4539337
#>  [5,] 0.3701371960 0.6298628
#>  [6,] 0.2143022565 0.7856977
#>  [7,] 0.1098832792 0.8901167
#>  [8,] 0.0523313209 0.9476687
#>  [9,] 0.0239814929 0.9760185
#> [10,] 0.0107891534 0.9892108
#> [11,] 0.0048130894 0.9951869
#> [12,] 0.0021389768 0.9978610
#> [13,] 0.0009489645 0.9990510
#> 
#> $prob.cats$CMC22
#>           resp.0    resp.1
#>  [1,] 0.70108252 0.2989175
#>  [2,] 0.65814446 0.3418555
#>  [3,] 0.60223077 0.3977692
#>  [4,] 0.53351931 0.4664807
#>  [5,] 0.45484828 0.5451517
#>  [6,] 0.37174912 0.6282509
#>  [7,] 0.29112113 0.7088789
#>  [8,] 0.21908747 0.7809125
#>  [9,] 0.15932508 0.8406749
#> [10,] 0.11271544 0.8872846
#> [11,] 0.07808592 0.9219141
#> [12,] 0.05327400 0.9467260
#> [13,] 0.03595484 0.9640452
#> 
#> $prob.cats$CMC23
#>           resp.0    resp.1
#>  [1,] 0.81417336 0.1858266
#>  [2,] 0.78255745 0.2174426
#>  [3,] 0.73597987 0.2640201
#>  [4,] 0.67090082 0.3290992
#>  [5,] 0.58639482 0.4136052
#>  [6,] 0.48652074 0.5134793
#>  [7,] 0.38081494 0.6191851
#>  [8,] 0.28123209 0.7187679
#>  [9,] 0.19719260 0.8028074
#> [10,] 0.13260986 0.8673901
#> [11,] 0.08646049 0.9135395
#> [12,] 0.05516987 0.9448301
#> [13,] 0.03470199 0.9652980
#> 
#> $prob.cats$CMC24
#>          resp.0    resp.1
#>  [1,] 0.7424407 0.2575593
#>  [2,] 0.7397318 0.2602682
#>  [3,] 0.7349855 0.2650145
#>  [4,] 0.7267326 0.2732674
#>  [5,] 0.7125710 0.2874290
#>  [6,] 0.6888139 0.3111861
#>  [7,] 0.6504326 0.3495674
#>  [8,] 0.5920444 0.4079556
#>  [9,] 0.5108752 0.4891248
#> [10,] 0.4111157 0.5888843
#> [11,] 0.3055225 0.6944775
#> [12,] 0.2100395 0.7899605
#> [13,] 0.1352352 0.8647648
#> 
#> $prob.cats$CMC25
#>           resp.0    resp.1
#>  [1,] 0.57491306 0.4250869
#>  [2,] 0.51959790 0.4804021
#>  [3,] 0.45453551 0.5454645
#>  [4,] 0.38315812 0.6168419
#>  [5,] 0.31059616 0.6894038
#>  [6,] 0.24232353 0.7576765
#>  [7,] 0.18260767 0.8173923
#>  [8,] 0.13362010 0.8663799
#>  [9,] 0.09550399 0.9044960
#> [10,] 0.06704943 0.9329506
#> [11,] 0.04645720 0.9535428
#> [12,] 0.03188754 0.9681125
#> [13,] 0.02174294 0.9782571
#> 
#> $prob.cats$CMC26
#>            resp.0    resp.1
#>  [1,] 0.649086399 0.3509136
#>  [2,] 0.559181392 0.4408186
#>  [3,] 0.451258810 0.5487412
#>  [4,] 0.338466291 0.6615337
#>  [5,] 0.236428722 0.7635713
#>  [6,] 0.155490483 0.8445095
#>  [7,] 0.097725330 0.9022747
#>  [8,] 0.059535733 0.9404643
#>  [9,] 0.035549017 0.9644510
#> [10,] 0.020964682 0.9790353
#> [11,] 0.012271697 0.9877283
#> [12,] 0.007151525 0.9928485
#> [13,] 0.004156851 0.9958431
#> 
#> $prob.cats$CMC27
#>           resp.0    resp.1
#>  [1,] 0.84513213 0.1548679
#>  [2,] 0.82768043 0.1723196
#>  [3,] 0.79805505 0.2019449
#>  [4,] 0.74980814 0.2501919
#>  [5,] 0.67630631 0.3236937
#>  [6,] 0.57498067 0.4250193
#>  [7,] 0.45297761 0.5470224
#>  [8,] 0.32790054 0.6720995
#>  [9,] 0.21912839 0.7808716
#> [10,] 0.13726929 0.8627307
#> [11,] 0.08212034 0.9178797
#> [12,] 0.04768227 0.9523177
#> [13,] 0.02718663 0.9728134
#> 
#> $prob.cats$CMC28
#>            resp.0    resp.1
#>  [1,] 0.790362846 0.2096372
#>  [2,] 0.787495672 0.2125043
#>  [3,] 0.779194161 0.2208058
#>  [4,] 0.755795178 0.2442048
#>  [5,] 0.694541338 0.3054587
#>  [6,] 0.561009869 0.4389901
#>  [7,] 0.358555891 0.6414441
#>  [8,] 0.174069895 0.8259301
#>  [9,] 0.069320439 0.9306796
#> [10,] 0.025049782 0.9749502
#> [11,] 0.008711171 0.9912888
#> [12,] 0.002987824 0.9970122
#> [13,] 0.001019889 0.9989801
#> 
#> $prob.cats$CMC29
#>            resp.0    resp.1
#>  [1,] 0.709594568 0.2904054
#>  [2,] 0.645008490 0.3549915
#>  [3,] 0.550016677 0.4499833
#>  [4,] 0.429885609 0.5701144
#>  [5,] 0.303926469 0.6960735
#>  [6,] 0.195328118 0.8046719
#>  [7,] 0.116404589 0.8835954
#>  [8,] 0.065888282 0.9341117
#>  [9,] 0.036133634 0.9638664
#> [10,] 0.019459154 0.9805408
#> [11,] 0.010374711 0.9896253
#> [12,] 0.005501369 0.9944986
#> [13,] 0.002908749 0.9970913
#> 
#> $prob.cats$CMC30
#>           resp.0    resp.1
#>  [1,] 0.67300598 0.3269940
#>  [2,] 0.66933473 0.3306653
#>  [3,] 0.66223751 0.3377625
#>  [4,] 0.64872356 0.3512764
#>  [5,] 0.62371805 0.3762819
#>  [6,] 0.57981037 0.4201896
#>  [7,] 0.50936395 0.4906361
#>  [8,] 0.41120865 0.5887913
#>  [9,] 0.29828301 0.7017170
#> [10,] 0.19375044 0.8062496
#> [11,] 0.11474733 0.8852527
#> [12,] 0.06371002 0.9362900
#> [13,] 0.03400001 0.9660000
#> 
#> $prob.cats$CMC31
#>          resp.0    resp.1
#>  [1,] 0.8753913 0.1246087
#>  [2,] 0.8556381 0.1443619
#>  [3,] 0.8274966 0.1725034
#>  [4,] 0.7884160 0.2115840
#>  [5,] 0.7360275 0.2639725
#>  [6,] 0.6690279 0.3309721
#>  [7,] 0.5883163 0.4116837
#>  [8,] 0.4978057 0.5021943
#>  [9,] 0.4041073 0.5958927
#> [10,] 0.3148052 0.6851948
#> [11,] 0.2361537 0.7638463
#> [12,] 0.1715559 0.8284441
#> [13,] 0.1214762 0.8785238
#> 
#> $prob.cats$CMC32
#>            resp.0    resp.1
#>  [1,] 0.713901229 0.2860988
#>  [2,] 0.683744966 0.3162550
#>  [3,] 0.627011415 0.3729886
#>  [4,] 0.532387892 0.4676121
#>  [5,] 0.402331244 0.5976688
#>  [6,] 0.264122628 0.7358774
#>  [7,] 0.152159421 0.8478406
#>  [8,] 0.079747338 0.9202527
#>  [9,] 0.039490946 0.9605091
#> [10,] 0.018974228 0.9810258
#> [11,] 0.008980408 0.9910196
#> [12,] 0.004219687 0.9957803
#> [13,] 0.001975936 0.9980241
#> 
#> $prob.cats$CMC33
#>            resp.0    resp.1
#>  [1,] 0.728479419 0.2715206
#>  [2,] 0.666409415 0.3335906
#>  [3,] 0.574056237 0.4259438
#>  [4,] 0.454999371 0.5450006
#>  [5,] 0.326963384 0.6730366
#>  [6,] 0.213517685 0.7864823
#>  [7,] 0.129004238 0.8709958
#>  [8,] 0.073828443 0.9261716
#>  [9,] 0.040844140 0.9591559
#> [10,] 0.022155024 0.9778450
#> [11,] 0.011886019 0.9881140
#> [12,] 0.006338659 0.9936613
#> [13,] 0.003369446 0.9966306
#> 
#> $prob.cats$CMC34
#>           resp.0    resp.1
#>  [1,] 0.82917132 0.1708287
#>  [2,] 0.81773286 0.1822671
#>  [3,] 0.79656973 0.2034303
#>  [4,] 0.75875154 0.2412485
#>  [5,] 0.69518745 0.3048125
#>  [6,] 0.59860745 0.4013925
#>  [7,] 0.47225247 0.5277475
#>  [8,] 0.33575924 0.6642408
#>  [9,] 0.21569569 0.7843043
#> [10,] 0.12773051 0.8722695
#> [11,] 0.07154086 0.9284591
#> [12,] 0.03872900 0.9612710
#> [13,] 0.02056441 0.9794356
#> 
#> $prob.cats$CMC35
#>            resp.0    resp.1
#>  [1,] 0.762211417 0.2377886
#>  [2,] 0.750189055 0.2498109
#>  [3,] 0.726337279 0.2736627
#>  [4,] 0.681243133 0.3187569
#>  [5,] 0.603267358 0.3967326
#>  [6,] 0.487173498 0.5128265
#>  [7,] 0.347819053 0.6521809
#>  [8,] 0.217995641 0.7820044
#>  [9,] 0.122670455 0.8773295
#> [10,] 0.064212229 0.9357878
#> [11,] 0.032232762 0.9677672
#> [12,] 0.015824706 0.9841753
#> [13,] 0.007682608 0.9923174
#> 
#> $prob.cats$CMC36
#>          resp.0    resp.1
#>  [1,] 0.8516403 0.1483597
#>  [2,] 0.8397174 0.1602826
#>  [3,] 0.8217479 0.1782521
#>  [4,] 0.7951695 0.2048305
#>  [5,] 0.7569296 0.2430704
#>  [6,] 0.7040453 0.2959547
#>  [7,] 0.6347730 0.3652270
#>  [8,] 0.5502140 0.4497860
#>  [9,] 0.4554486 0.5445514
#> [10,] 0.3588973 0.6411027
#> [11,] 0.2696201 0.7303799
#> [12,] 0.1941748 0.8058252
#> [13,] 0.1351187 0.8648813
#> 
#> $prob.cats$CMC37
#>            resp.0    resp.1
#>  [1,] 0.885406220 0.1145938
#>  [2,] 0.871965346 0.1280347
#>  [3,] 0.841639854 0.1583601
#>  [4,] 0.777462104 0.2225379
#>  [5,] 0.658313824 0.3416862
#>  [6,] 0.482722603 0.5172774
#>  [7,] 0.295588227 0.7044118
#>  [8,] 0.153941207 0.8460588
#>  [9,] 0.072022321 0.9279777
#> [10,] 0.031825120 0.9681749
#> [11,] 0.013689850 0.9863101
#> [12,] 0.005819175 0.9941808
#> [13,] 0.002460940 0.9975391
#> 
#> $prob.cats$CMC38
#>           resp.0    resp.1
#>  [1,] 0.68162335 0.3183766
#>  [2,] 0.64525762 0.3547424
#>  [3,] 0.59914839 0.4008516
#>  [4,] 0.54316115 0.4568388
#>  [5,] 0.47864430 0.5213557
#>  [6,] 0.40863088 0.5913691
#>  [7,] 0.33743374 0.6625663
#>  [8,] 0.26966099 0.7303390
#>  [9,] 0.20908558 0.7909144
#> [10,] 0.15791617 0.8420838
#> [11,] 0.11671436 0.8832856
#> [12,] 0.08479973 0.9152003
#> [13,] 0.06081314 0.9391869
#> 
#> $prob.cats$CFR1
#>             resp.0       resp.1       resp.2      resp.3      resp.4
#>  [1,] 8.968354e-01 0.0699079775 0.0208141796 0.007476110 0.004966303
#>  [2,] 7.695561e-01 0.1482526323 0.0504347170 0.018930447 0.012826149
#>  [3,] 5.619465e-01 0.2490041213 0.1103856547 0.045947505 0.032716269
#>  [4,] 3.301127e-01 0.2922192093 0.1958237422 0.100921850 0.080922492
#>  [5,] 1.591701e-01 0.2284605147 0.2458460313 0.180057239 0.186466137
#>  [6,] 6.778915e-02 0.1278110454 0.2034126488 0.227292069 0.373695086
#>  [7,] 2.717524e-02 0.0582541544 0.1177855461 0.188444156 0.608340901
#>  [8,] 1.061686e-02 0.0240227004 0.0545913855 0.109048243 0.801720811
#>  [9,] 4.105233e-03 0.0094913635 0.0226740334 0.050491583 0.913237787
#> [10,] 1.580993e-03 0.0036861472 0.0089843055 0.020959046 0.964789508
#> [11,] 6.079191e-04 0.0014219999 0.0034931473 0.008302597 0.986174336
#> [12,] 2.336153e-04 0.0005471397 0.0013481328 0.003227752 0.994643360
#> [13,] 8.975455e-05 0.0002103112 0.0005188049 0.001245655 0.997935474
#> 
#> $prob.cats$CFR2
#>            resp.0     resp.1     resp.2      resp.3     resp.4
#>  [1,] 0.948323850 0.02864214 0.01270891 0.004875763 0.00544934
#>  [2,] 0.906401398 0.05083025 0.02338117 0.009110661 0.01027652
#>  [3,] 0.836338601 0.08560225 0.04194655 0.016815806 0.01929680
#>  [4,] 0.729482976 0.13225270 0.07197265 0.030344585 0.03594709
#>  [5,] 0.587288675 0.17955046 0.11457255 0.052590990 0.06599733
#>  [6,] 0.428869562 0.20557118 0.16239502 0.085073103 0.11809113
#>  [7,] 0.283798372 0.19423642 0.19619988 0.123371440 0.20239388
#>  [8,] 0.172940276 0.15288011 0.19620737 0.153252715 0.32471953
#>  [9,] 0.099377228 0.10382760 0.16241258 0.157599240 0.47678336
#> [10,] 0.055023671 0.06359086 0.11459117 0.133517682 0.63327662
#> [11,] 0.029810541 0.03649666 0.07198700 0.095764713 0.76594109
#> [12,] 0.015955584 0.02016575 0.04195586 0.060785863 0.86113694
#> [13,] 0.008483646 0.01090830 0.02338666 0.035642282 0.92157910
#> 
#> $prob.cats$AMC1
#>           resp.0    resp.1
#>  [1,] 0.77089868 0.2291013
#>  [2,] 0.76692599 0.2330740
#>  [3,] 0.75878930 0.2412107
#>  [4,] 0.74239956 0.2576004
#>  [5,] 0.71046694 0.2895331
#>  [6,] 0.65210197 0.3478980
#>  [7,] 0.55690382 0.4430962
#>  [8,] 0.42714698 0.5728530
#>  [9,] 0.28767390 0.7123261
#> [10,] 0.17129127 0.8287087
#> [11,] 0.09300027 0.9069997
#> [12,] 0.04766607 0.9523339
#> [13,] 0.02366342 0.9763366
#> 
#> $prob.cats$AMC2
#>             resp.0    resp.1
#>  [1,] 0.7824902694 0.2175097
#>  [2,] 0.7123929195 0.2876071
#>  [3,] 0.5858124437 0.4141876
#>  [4,] 0.4100660501 0.5899339
#>  [5,] 0.2379091459 0.7620909
#>  [6,] 0.1182073607 0.8817926
#>  [7,] 0.0534005468 0.9465995
#>  [8,] 0.0229937270 0.9770063
#>  [9,] 0.0096879188 0.9903121
#> [10,] 0.0040437409 0.9959563
#> [11,] 0.0016812097 0.9983188
#> [12,] 0.0006978224 0.9993022
#> [13,] 0.0002894479 0.9997106
#> 
#> $prob.cats$AMC3
#>           resp.0    resp.1
#>  [1,] 0.85600400 0.1439960
#>  [2,] 0.85060362 0.1493964
#>  [3,] 0.83970377 0.1602962
#>  [4,] 0.81813309 0.1818669
#>  [5,] 0.77706170 0.2229383
#>  [6,] 0.70431268 0.2956873
#>  [7,] 0.59055555 0.4094444
#>  [8,] 0.44325280 0.5567472
#>  [9,] 0.29292193 0.7070781
#> [10,] 0.17252834 0.8274717
#> [11,] 0.09347767 0.9065223
#> [12,] 0.04811644 0.9518836
#> [13,] 0.02407696 0.9759230
#> 
#> $prob.cats$AMC4
#>           resp.0    resp.1
#>  [1,] 0.81139437 0.1886056
#>  [2,] 0.77678254 0.2232175
#>  [3,] 0.72625330 0.2737467
#>  [4,] 0.65659064 0.3434094
#>  [5,] 0.56776425 0.4322358
#>  [6,] 0.46513165 0.5348683
#>  [7,] 0.35922073 0.6407793
#>  [8,] 0.26193969 0.7380603
#>  [9,] 0.18168787 0.8183121
#> [10,] 0.12114890 0.8788511
#> [11,] 0.07849153 0.9215085
#> [12,] 0.04985814 0.9501419
#> [13,] 0.03125916 0.9687408
#> 
#> $prob.cats$AMC5
#>          resp.0    resp.1
#>  [1,] 0.8311330 0.1688670
#>  [2,] 0.8285577 0.1714423
#>  [3,] 0.8243643 0.1756357
#>  [4,] 0.8175709 0.1824291
#>  [5,] 0.8066573 0.1933427
#>  [6,] 0.7893581 0.2106419
#>  [7,] 0.7625113 0.2374887
#>  [8,] 0.7221861 0.2778139
#>  [9,] 0.6644921 0.3355079
#> [10,] 0.5874447 0.4125553
#> [11,] 0.4935013 0.5064987
#> [12,] 0.3908772 0.6091228
#> [13,] 0.2913936 0.7086064
#> 
#> $prob.cats$AMC6
#>           resp.0    resp.1
#>  [1,] 0.82156825 0.1784317
#>  [2,] 0.82152096 0.1784790
#>  [3,] 0.82137363 0.1786264
#>  [4,] 0.82091484 0.1790852
#>  [5,] 0.81948840 0.1805116
#>  [6,] 0.81507487 0.1849251
#>  [7,] 0.80162101 0.1983790
#>  [8,] 0.76240442 0.2375956
#>  [9,] 0.66154935 0.3384506
#> [10,] 0.46844069 0.5315593
#> [11,] 0.24530186 0.7546981
#> [12,] 0.09873626 0.9012637
#> [13,] 0.03449975 0.9655002
#> 
#> $prob.cats$AMC7
#>           resp.0    resp.1
#>  [1,] 0.84533340 0.1546666
#>  [2,] 0.82654527 0.1734547
#>  [3,] 0.79397527 0.2060247
#>  [4,] 0.74014543 0.2598546
#>  [5,] 0.65782939 0.3421706
#>  [6,] 0.54579486 0.4542051
#>  [7,] 0.41526291 0.5847371
#>  [8,] 0.28809487 0.7119051
#>  [9,] 0.18406148 0.8159385
#> [10,] 0.11044873 0.8895513
#> [11,] 0.06354381 0.9364562
#> [12,] 0.03562245 0.9643775
#> [13,] 0.01967010 0.9803299
#> 
#> $prob.cats$AMC8
#>            resp.0    resp.1
#>  [1,] 0.812809384 0.1871906
#>  [2,] 0.807132544 0.1928675
#>  [3,] 0.794533320 0.2054667
#>  [4,] 0.767337684 0.2326623
#>  [5,] 0.712004746 0.2879953
#>  [6,] 0.611833276 0.3881667
#>  [7,] 0.463710502 0.5362895
#>  [8,] 0.299231663 0.7007683
#>  [9,] 0.165748506 0.8342515
#> [10,] 0.082346375 0.9176536
#> [11,] 0.038435419 0.9615646
#> [12,] 0.017385962 0.9826140
#> [13,] 0.007749707 0.9922503
#> 
#> $prob.cats$AMC9
#>           resp.0     resp.1
#>  [1,] 0.90070981 0.09929019
#>  [2,] 0.88634918 0.11365082
#>  [3,] 0.86121610 0.13878390
#>  [4,] 0.81872262 0.18127738
#>  [5,] 0.75090538 0.24909462
#>  [6,] 0.65204433 0.34795567
#>  [7,] 0.52542606 0.47457394
#>  [8,] 0.38764297 0.61235703
#>  [9,] 0.26192331 0.73807669
#> [10,] 0.16435638 0.83564362
#> [11,] 0.09772499 0.90227501
#> [12,] 0.05609748 0.94390252
#> [13,] 0.03152101 0.96847899
#> 
#> $prob.cats$AMC10
#>           resp.0     resp.1
#>  [1,] 0.91769489 0.08230511
#>  [2,] 0.91492491 0.08507509
#>  [3,] 0.90961229 0.09038771
#>  [4,] 0.89950576 0.10049424
#>  [5,] 0.88057421 0.11942579
#>  [6,] 0.84611646 0.15388354
#>  [7,] 0.78656018 0.21343982
#>  [8,] 0.69227364 0.30772636
#>  [9,] 0.56193373 0.43806627
#> [10,] 0.41188049 0.58811951
#> [11,] 0.27156436 0.72843564
#> [12,] 0.16367280 0.83632720
#> [13,] 0.09253548 0.90746452
#> 
#> $prob.cats$AMC11
#>             resp.0    resp.1
#>  [1,] 0.7306673001 0.2693327
#>  [2,] 0.7015938940 0.2984061
#>  [3,] 0.6408168868 0.3591831
#>  [4,] 0.5311449135 0.4688551
#>  [5,] 0.3772510537 0.6227489
#>  [6,] 0.2231439827 0.7768560
#>  [7,] 0.1130592200 0.8869408
#>  [8,] 0.0519597008 0.9480403
#>  [9,] 0.0227083329 0.9772917
#> [10,] 0.0096966895 0.9903033
#> [11,] 0.0040987541 0.9959012
#> [12,] 0.0017250302 0.9982750
#> [13,] 0.0007246784 0.9992753
#> 
#> $prob.cats$AMC12
#>           resp.0    resp.1
#>  [1,] 0.69999529 0.3000047
#>  [2,] 0.65922677 0.3407732
#>  [3,] 0.60213490 0.3978651
#>  [4,] 0.52772993 0.4722701
#>  [5,] 0.43934732 0.5606527
#>  [6,] 0.34522237 0.6547776
#>  [7,] 0.25595220 0.7440478
#>  [8,] 0.18012413 0.8198759
#>  [9,] 0.12151623 0.8784838
#> [10,] 0.07943763 0.9205624
#> [11,] 0.05079981 0.9492002
#> [12,] 0.03201198 0.9679880
#> [13,] 0.01998132 0.9800187
#> 
#> $prob.cats$AFR1
#>           resp.0     resp.1     resp.2      resp.3      resp.4
#>  [1,] 0.95202932 0.02285370 0.01273579 0.005599951 0.006781242
#>  [2,] 0.91827043 0.03820143 0.02186387 0.009747982 0.011916292
#>  [3,] 0.86414428 0.06145074 0.03676271 0.016784107 0.020858164
#>  [4,] 0.78265679 0.09300604 0.05971112 0.028362379 0.036263669
#>  [5,] 0.67090626 0.12857484 0.09175201 0.046443687 0.062323199
#>  [6,] 0.53577801 0.15720974 0.12967155 0.072272883 0.105067824
#>  [7,] 0.39518402 0.16580943 0.16323610 0.104011483 0.171758963
#>  [8,] 0.27002368 0.14974213 0.17810706 0.134026838 0.268100294
#>  [9,] 0.17315470 0.11740555 0.16646606 0.150124782 0.392848907
#> [10,] 0.10599097 0.08223262 0.13450735 0.143924338 0.533344722
#> [11,] 0.06289718 0.05313797 0.09642167 0.118799825 0.668743350
#> [12,] 0.03660702 0.03256656 0.06331765 0.086520189 0.780988580
#> [13,] 0.02105884 0.01931415 0.03920909 0.057425886 0.862992035
#> 
#> $prob.cats$AFR2
#>            resp.0      resp.1      resp.2     resp.3     resp.4
#>  [1,] 0.756996282 0.127731524 0.059548711 0.02798118 0.02774230
#>  [2,] 0.627052778 0.178488730 0.095898636 0.04834737 0.05021248
#>  [3,] 0.475745551 0.215213804 0.140589151 0.07923886 0.08921263
#>  [4,] 0.328763740 0.218078592 0.180256757 0.11929555 0.15360536
#>  [5,] 0.209081698 0.185339005 0.195410155 0.15853414 0.25163500
#>  [6,] 0.124863646 0.135235136 0.176882725 0.17916490 0.38385359
#>  [7,] 0.071501841 0.087972693 0.135757004 0.16896375 0.53580471
#>  [8,] 0.039904908 0.052986518 0.091511511 0.13421029 0.68138678
#>  [9,] 0.021940832 0.030434708 0.056383019 0.09275843 0.79848301
#> [10,] 0.011962917 0.017003993 0.032826537 0.05809069 0.88011586
#> [11,] 0.006492484 0.009353065 0.018482527 0.03415577 0.93151615
#> [12,] 0.003514686 0.005100451 0.010210170 0.01934045 0.96183424
#> [13,] 0.001900052 0.002768300 0.005581024 0.01071812 0.97903250
#> 
#> $prob.cats$AFR3
#>           resp.0     resp.1     resp.2     resp.3     resp.4
#>  [1,] 0.87728238 0.05507504 0.02890256 0.01511871 0.02362130
#>  [2,] 0.82183517 0.07709251 0.04228592 0.02264791 0.03613849
#>  [3,] 0.74851698 0.10308914 0.06014028 0.03333801 0.05491559
#>  [4,] 0.65759750 0.12977199 0.08218608 0.04783159 0.08261283
#>  [5,] 0.55341880 0.15154105 0.10640800 0.06616293 0.12246922
#>  [6,] 0.44432718 0.16224285 0.12856040 0.08704262 0.17782695
#>  [7,] 0.34035092 0.15835028 0.14298680 0.10726343 0.25104857
#>  [8,] 0.24976929 0.14118385 0.14513047 0.12203367 0.34188272
#>  [9,] 0.17683270 0.11605095 0.13425709 0.12684872 0.44601053
#> [10,] 0.12173865 0.08915750 0.11393906 0.12006221 0.55510258
#> [11,] 0.08209777 0.06498784 0.08981437 0.10396960 0.65913042
#> [12,] 0.05456295 0.04556924 0.06675328 0.08331681 0.74979773
#> [13,] 0.03590184 0.03108838 0.04746942 0.06270945 0.82283090
#> 
#> 
#> $icc
#>            CMC1      CMC2      CMC3      CMC4      CMC5       CMC6      CMC7
#>  [1,] 0.2852746 0.1943080 0.1368890 0.2501262 0.2246487 0.07414683 0.1420085
#>  [2,] 0.2957986 0.2232020 0.1573778 0.2806501 0.2555248 0.07677376 0.1544198
#>  [3,] 0.3106618 0.2897664 0.1880131 0.3270196 0.2985985 0.08284878 0.1732339
#>  [4,] 0.3313283 0.4197163 0.2323701 0.3932257 0.3560543 0.09672461 0.2011940
#>  [5,] 0.3594488 0.6074293 0.2936937 0.4798436 0.4282772 0.12753967 0.2415448
#>  [6,] 0.3966064 0.7872691 0.3732690 0.5810712 0.5125815 0.19189848 0.2973789
#>  [7,] 0.4438479 0.9031813 0.4684502 0.6848773 0.6028847 0.31061786 0.3703025
#>  [8,] 0.5010506 0.9600577 0.5718055 0.7779937 0.6911312 0.48667312 0.4586846
#>  [9,] 0.5663616 0.9842451 0.6728935 0.8520134 0.7699629 0.67840589 0.5565897
#> [10,] 0.6361104 0.9938995 0.7621415 0.9054015 0.8349326 0.82820554 0.6548062
#> [11,] 0.7054788 0.9976550 0.8340879 0.9412673 0.8850143 0.91770056 0.7440078
#> [12,] 0.7697514 0.9991011 0.8879534 0.9642258 0.9216658 0.96286386 0.8179954
#> [13,] 0.8255104 0.9996558 0.9260909 0.9784710 0.9474814 0.98372018 0.8748828
#>            CMC8      CMC9     CMC10     CMC11     CMC12     CMC13     CMC14
#>  [1,] 0.1488735 0.2819889 0.1429166 0.1956330 0.1102839 0.1826100 0.2540700
#>  [2,] 0.1658894 0.2954650 0.1514253 0.2320279 0.1222729 0.1840354 0.2607542
#>  [3,] 0.1905709 0.3151658 0.1691683 0.2853869 0.1402474 0.1869725 0.2746135
#>  [4,] 0.2255244 0.3433112 0.2049298 0.3588958 0.1667225 0.1929903 0.3024954
#>  [5,] 0.2733819 0.3822262 0.2723085 0.4519323 0.2047183 0.2051780 0.3553294
#>  [6,] 0.3359664 0.4336656 0.3845127 0.5578255 0.2572631 0.2292911 0.4449700
#>  [7,] 0.4130708 0.4977632 0.5378155 0.6647267 0.3263209 0.2748651 0.5718934
#>  [8,] 0.5013776 0.5720086 0.6990598 0.7603424 0.4112502 0.3540015 0.7122130
#>  [9,] 0.5944440 0.6510701 0.8280817 0.8370376 0.5075670 0.4731215 0.8305997
#> [10,] 0.6843131 0.7280648 0.9105648 0.8933571 0.6072530 0.6187582 0.9097547
#> [11,] 0.7640574 0.7967921 0.9559885 0.9320885 0.7011234 0.7575777 0.9547936
#> [12,] 0.8296788 0.8535305 0.9789676 0.9575381 0.7819728 0.8618156 0.9780988
#> [13,] 0.8804093 0.8974274 0.9900938 0.9737635 0.8464268 0.9269167 0.9895669
#>           CMC15      CMC16     CMC17     CMC18     CMC19     CMC20     CMC21
#>  [1,] 0.1357108 0.05331770 0.1932171 0.2736481 0.1878184 0.2915193 0.1660267
#>  [2,] 0.1563288 0.05761487 0.2075524 0.2743693 0.2050723 0.3744156 0.2155500
#>  [3,] 0.1931363 0.06937043 0.2348989 0.2760474 0.2553291 0.4960884 0.3082638
#>  [4,] 0.2552674 0.10059246 0.2845622 0.2799351 0.3798782 0.6404138 0.4539337
#>  [5,] 0.3508543 0.17739327 0.3671895 0.2888510 0.5950464 0.7739936 0.6298628
#>  [6,] 0.4787818 0.33533633 0.4864879 0.3088309 0.8073754 0.8720672 0.7856977
#>  [7,] 0.6214416 0.56680113 0.6276798 0.3513755 0.9277400 0.9325069 0.8901167
#>  [8,] 0.7515033 0.78000253 0.7602706 0.4328754 0.9757931 0.9658250 0.9476687
#>  [9,] 0.8500319 0.90715819 0.8603266 0.5615063 0.9922230 0.9830709 0.9760185
#> [10,] 0.9147040 0.96440728 0.9240726 0.7133134 0.9975360 0.9917071 0.9892108
#> [11,] 0.9532533 0.98689511 0.9604145 0.8414565 0.9992228 0.9959600 0.9951869
#> [12,] 0.9749245 0.99524881 0.9798322 0.9224170 0.9997552 0.9980373 0.9978610
#> [13,] 0.9867079 0.99828719 0.9898487 0.9646092 0.9999229 0.9990477 0.9990510
#>           CMC22     CMC23     CMC24     CMC25     CMC26     CMC27     CMC28
#>  [1,] 0.2989175 0.1858266 0.2575593 0.4250869 0.3509136 0.1548679 0.2096372
#>  [2,] 0.3418555 0.2174426 0.2602682 0.4804021 0.4408186 0.1723196 0.2125043
#>  [3,] 0.3977692 0.2640201 0.2650145 0.5454645 0.5487412 0.2019449 0.2208058
#>  [4,] 0.4664807 0.3290992 0.2732674 0.6168419 0.6615337 0.2501919 0.2442048
#>  [5,] 0.5451517 0.4136052 0.2874290 0.6894038 0.7635713 0.3236937 0.3054587
#>  [6,] 0.6282509 0.5134793 0.3111861 0.7576765 0.8445095 0.4250193 0.4389901
#>  [7,] 0.7088789 0.6191851 0.3495674 0.8173923 0.9022747 0.5470224 0.6414441
#>  [8,] 0.7809125 0.7187679 0.4079556 0.8663799 0.9404643 0.6720995 0.8259301
#>  [9,] 0.8406749 0.8028074 0.4891248 0.9044960 0.9644510 0.7808716 0.9306796
#> [10,] 0.8872846 0.8673901 0.5888843 0.9329506 0.9790353 0.8627307 0.9749502
#> [11,] 0.9219141 0.9135395 0.6944775 0.9535428 0.9877283 0.9178797 0.9912888
#> [12,] 0.9467260 0.9448301 0.7899605 0.9681125 0.9928485 0.9523177 0.9970122
#> [13,] 0.9640452 0.9652980 0.8647648 0.9782571 0.9958431 0.9728134 0.9989801
#>           CMC29     CMC30     CMC31     CMC32     CMC33     CMC34     CMC35
#>  [1,] 0.2904054 0.3269940 0.1246087 0.2860988 0.2715206 0.1708287 0.2377886
#>  [2,] 0.3549915 0.3306653 0.1443619 0.3162550 0.3335906 0.1822671 0.2498109
#>  [3,] 0.4499833 0.3377625 0.1725034 0.3729886 0.4259438 0.2034303 0.2736627
#>  [4,] 0.5701144 0.3512764 0.2115840 0.4676121 0.5450006 0.2412485 0.3187569
#>  [5,] 0.6960735 0.3762819 0.2639725 0.5976688 0.6730366 0.3048125 0.3967326
#>  [6,] 0.8046719 0.4201896 0.3309721 0.7358774 0.7864823 0.4013925 0.5128265
#>  [7,] 0.8835954 0.4906361 0.4116837 0.8478406 0.8709958 0.5277475 0.6521809
#>  [8,] 0.9341117 0.5887913 0.5021943 0.9202527 0.9261716 0.6642408 0.7820044
#>  [9,] 0.9638664 0.7017170 0.5958927 0.9605091 0.9591559 0.7843043 0.8773295
#> [10,] 0.9805408 0.8062496 0.6851948 0.9810258 0.9778450 0.8722695 0.9357878
#> [11,] 0.9896253 0.8852527 0.7638463 0.9910196 0.9881140 0.9284591 0.9677672
#> [12,] 0.9944986 0.9362900 0.8284441 0.9957803 0.9936613 0.9612710 0.9841753
#> [13,] 0.9970913 0.9660000 0.8785238 0.9980241 0.9966306 0.9794356 0.9923174
#>           CMC36     CMC37     CMC38      CFR1      CFR2      AMC1      AMC2
#>  [1,] 0.1483597 0.1145938 0.3183766 0.1538299 0.0904846 0.2291013 0.2175097
#>  [2,] 0.1602826 0.1280347 0.3547424 0.3572180 0.1660307 0.2330740 0.2876071
#>  [3,] 0.1782521 0.1583601 0.4008516 0.7384830 0.2971299 0.2412107 0.4141876
#>  [4,] 0.2048305 0.2225379 0.4568388 1.3103222 0.5110201 0.2576004 0.5899339
#>  [5,] 0.2430704 0.3416862 0.5213557 2.0061888 0.8304578 0.2895331 0.7620909
#>  [6,] 0.2959547 0.5172774 0.5913691 2.7112929 1.2579451 0.3478980 0.8817926
#>  [7,] 0.3652270 0.7044118 0.6625663 3.2925213 1.7663260 0.4430962 0.9465995
#>  [8,] 0.4497860 0.8460588 0.7303390 3.6672334 2.3039311 0.5728530 0.9770063
#>  [9,] 0.5445514 0.9279777 0.7909144 3.8592653 2.8085839 0.7123261 0.9903121
#> [10,] 0.6411027 0.9681749 0.8420838 3.9436899 3.2264327 0.8287087 0.9959563
#> [11,] 0.7303799 0.9863101 0.8832856 3.9780134 3.5315291 0.9069997 0.9983188
#> [12,] 0.8058252 0.9941808 0.9152003 3.9915001 3.7309828 0.9523339 0.9993022
#> [13,] 0.8648813 0.9975391 0.9391869 3.9967268 3.8509249 0.9763366 0.9997106
#>            AMC3      AMC4      AMC5      AMC6      AMC7      AMC8       AMC9
#>  [1,] 0.1439960 0.1886056 0.1688670 0.1784317 0.1546666 0.1871906 0.09929019
#>  [2,] 0.1493964 0.2232175 0.1714423 0.1784790 0.1734547 0.1928675 0.11365082
#>  [3,] 0.1602962 0.2737467 0.1756357 0.1786264 0.2060247 0.2054667 0.13878390
#>  [4,] 0.1818669 0.3434094 0.1824291 0.1790852 0.2598546 0.2326623 0.18127738
#>  [5,] 0.2229383 0.4322358 0.1933427 0.1805116 0.3421706 0.2879953 0.24909462
#>  [6,] 0.2956873 0.5348683 0.2106419 0.1849251 0.4542051 0.3881667 0.34795567
#>  [7,] 0.4094444 0.6407793 0.2374887 0.1983790 0.5847371 0.5362895 0.47457394
#>  [8,] 0.5567472 0.7380603 0.2778139 0.2375956 0.7119051 0.7007683 0.61235703
#>  [9,] 0.7070781 0.8183121 0.3355079 0.3384506 0.8159385 0.8342515 0.73807669
#> [10,] 0.8274717 0.8788511 0.4125553 0.5315593 0.8895513 0.9176536 0.83564362
#> [11,] 0.9065223 0.9215085 0.5064987 0.7546981 0.9364562 0.9615646 0.90227501
#> [12,] 0.9518836 0.9501419 0.6091228 0.9012637 0.9643775 0.9826140 0.94390252
#> [13,] 0.9759230 0.9687408 0.7086064 0.9655002 0.9803299 0.9922503 0.96847899
#>            AMC10     AMC11     AMC12      AFR1      AFR2      AFR3
#>  [1,] 0.08230511 0.2693327 0.3000047 0.0922501 0.4417417 0.2527215
#>  [2,] 0.08507509 0.2984061 0.3407732 0.1588383 0.7161781 0.3741620
#>  [3,] 0.09038771 0.3591831 0.3978651 0.2687611 1.0909592 0.5430461
#>  [4,] 0.10049424 0.4688551 0.4722701 0.4425701 1.5509002 0.7680903
#>  [5,] 0.11942579 0.6227489 0.5606527 0.7007027 2.0583017 1.0527227
#>  [6,] 0.15388354 0.7768560 0.6547776 1.0536428 2.5619097 1.3917993
#>  [7,] 0.21343982 0.8869408 0.7440478 1.4913519 3.0095968 1.7703084
#>  [8,] 0.30772636 0.9480403 0.8198759 1.9804379 3.3641875 2.1650767
#>  [9,] 0.43806627 0.9772917 0.8784838 2.4721076 3.6154081 2.5491534
#> [10,] 0.58811951 0.9903033 0.9205624 2.9163992 3.7773926 2.8976326
#> [11,] 0.72843564 0.9959012 0.9492002 3.2773542 3.8748500 3.1930471
#> [12,] 0.83632720 0.9982750 0.9679880 3.5427168 3.9308791 3.4282171
#> [13,] 0.90746452 0.9992753 0.9800187 3.7219781 3.9622147 3.6054792
#> 
#> $tcc
#>  [1] 11.34775 13.24700 16.26269 20.68413 26.52016 33.38252 40.63685 47.56966
#>  [9] 53.64067 58.62450 62.43763 65.11620 66.88748
#> 
#> $theta
#>  [1] -3.0 -2.5 -2.0 -1.5 -1.0 -0.5  0.0  0.5  1.0  1.5  2.0  2.5  3.0
#> 
#> attr(,"class")
#> [1] "traceline"